% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{avgNormDiff}
\alias{avgNormDiff}
\title{takes average of normdiff of each row in x}
\usage{
avgNormDiff(x)
}
\arguments{
\item{x}{(numeric) matrix of values, one column per patient (e.g. ages)}
}
\value{
symmetric matrix of size ncol(dat) (number of patients) containing
pairwise patient similarities
}
\description{
takes average of normdiff of each row in x
}
\examples{
data(xpr)
sim <- avgNormDiff(xpr[,seq_len(2)])
}
