% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPredictor.R
\name{buildPredictor}
\alias{buildPredictor}
\title{Run nested cross-validation on data}
\usage{
buildPredictor(pheno, dataList, groupList, outDir, makeNetFunc,
  featScoreMax = 10L, trainProp = 0.8, numSplits = 10L, numCores,
  JavaMemory = 4L, featSelCutoff = 9L, keepAllData = FALSE,
  startAt = 1L, preFilter = FALSE, impute = FALSE,
  logging = "default")
}
\arguments{
\item{pheno}{(data.frame) sample metadata, must have ID and STATUS columns}

\item{dataList}{(list) keys are datatypes; values contain patient data
for the corresponding datatype. e.g. dataList[["rna"]] contains expression
matrix. Rows are units (e.g. genes, individual clinical variables) and 
columns are patients}

\item{groupList}{(list) keys are datatypes and values are lists indicating
how units for those datatypes are to be grouped. Keys must match those 
in dataList. Each entry of groupList[[k]] will generate a new PSN.
 e.g. groupList[["rna"]] could be a list of pathway definitions. 
So keys(groupList[["rna"]]) would have pathway names, generating one PSN
per pathways, and values(groupList[["rna"]]) would be genes that would be
grouped for the corresponding pathwayList.}

\item{outDir}{(char) directory where results will be stored. If this 
directory exists, its contents will be overwritten}

\item{makeNetFunc}{(function) user-defined function for creating the set
of input PSN provided to netDx. See createPSN_MultiData()::customFunc.}

\item{featScoreMax}{(integer) number of CV folds in inner loop}

\item{trainProp}{(numeric 0 to 1) Percent samples to use for training}

\item{numSplits}{(integer) number of train/blind test splits (i.e. iterations
of outer loop)}

\item{numCores}{(integer) number of CPU cores for parallel processing}

\item{JavaMemory}{(integer) memory in (Gb) used for each fold of CV}

\item{featSelCutoff}{(integer) cutoff for inner-fold CV to call feature-selected
in a given split}

\item{keepAllData}{(logical) if TRUE keeps all intermediate files, even
those not needed for assessing the predictor. Use very cautiously as for
some designs, each split can result in using 1Gb of data.}

\item{startAt}{(integer) which of the splits to start at (e.g. if the
job aborted part-way through)}

\item{preFilter}{(logical) if TRUE uses lasso to prefilter dataList within 
cross-validation loop. Only variables that pass lasso get included. The
current option is not recommended for pathway-level features as most genes
will be eliminated by lasso. Future variations may allow other prefiltering
options that are more lenient.}

\item{impute}{(logical) if TRUE applies imputation by median within CV}

\item{logging}{(char) level of detail with which messages are printed. Options are: 1) none: turn off all messages; 2) all: greatest level of detail (recommended for advanced users, or for debugging); 3) default: print key details (useful setting for most users)}
}
\description{
Run nested cross-validation on data
}
\details{
wrapper function to run netDx with nested cross-validation, 
with an inner loop of X-fold cross-validation and an outer loop of different
random splits of data into train and blind test. The user needs to supply
a custom function to create PSN, see createPSN_MultiData(). This wrapper
provides flexibility for designs with one or several heterogeneous data
types, and one or more ways of defining patient similarity. 
For example, designs it handles includes
1) Single datatype, single similarity metric: Expression data -> pathways
2) Single datatype, multiple metrics: Expression data -> pathways
(Pearson corr) and single gene networks (normalized difference)
3) Multiple datatypes, multiple metrics: Expression -> Pathways; 
Clinical -> single or grouped nets
}
