% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callFeatSel.R
\name{callFeatSel}
\alias{callFeatSel}
\title{Return feature selected nets based on given criteria}
\usage{
callFeatSel(netScores, fsCutoff, fsPctPass)
}
\arguments{
\item{netScores}{(matrix) matrix of net scores}

\item{fsCutoff}{(integer) net must score at least this much in a split to
"pass" the threshold}

\item{fsPctPass}{(numeric 0 to 1) net must pass at least this percent of
splits to be considered feature-selected}
}
\value{
(char) names of nets that pass feature-selection
}
\description{
Return feature selected nets based on given criteria
}
\details{
given the output of genNetScores.R and criteria for defining
feature-selected (FS) nets, returns subset of nets that pass criteria.
Net must score <fsCutoff> for at least <fsPctPass> % of splits, to be
considered feature-selected.
}
\examples{
data(featScores)
passed <- lapply(featScores, function(x) {
   callFeatSel(x,10,0.7) # score 10/10 in >=70\% of trials
})
print(passed)
}
