% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{callOverallSelectedFeatures}
\alias{callOverallSelectedFeatures}
\title{Wrapper to call selected features}
\usage{
callOverallSelectedFeatures(
  featScores,
  featureSelCutoff,
  featureSelPct,
  cleanNames = TRUE
)
}
\arguments{
\item{featScores}{(list of lists): matrix of feature scores across all splits, separated
by patient label. First level: patient labels. Second level: matrix of scores for 
corresponding label.}

\item{featureSelCutoff}{(integer) cutoff score for feature selection.
A feature must have minimum of this score for specified fraction of splits 
(see featureSelPct) to pass.}

\item{featureSelPct}{(numeric between 0 and 1) cutoff percent for feature selection.
A feature must have minimum score of featureSelCutoff for featureSelPct of 
train/test splits to pass.}

\item{cleanNames}{(logical) remove internal suffixes for human readability}
}
\value{
(list) Feature scores for all splits, plus those passing selection for overall predictor
featScores: (matrix) feature scores for each split
selectedFeatures: (list) features passing selection for each class; one key per class
}
\description{
Wrapper to call selected features
}
\details{
Calls features that are consistently high-scoring for predicting 
each class. The context for this is as follows: 
The original model runs feature selection over multiple splits of data
into train/test samples, and each such split generates scores for all features.
This function identifies features with scores that exceed a threshold for a fraction
of train/test splits; the threshold and fraction are both user-specified. This
function is called by the wrapper getResults(), which returns both the matrix of 
feature scores across splits and list of features that pass the user-specified cutoffs.
}
