% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPathwayName.R
\name{cleanPathwayName}
\alias{cleanPathwayName}
\title{Clean pathway name so it can be a filename.}
\usage{
cleanPathwayName(curP)
}
\arguments{
\item{curP}{(char) pathway name}
}
\value{
(char) Cleaned pathway name
}
\description{
Clean pathway name so it can be a filename.
}
\examples{
cleanPathwayName("7-(3-AMINO-3-CARBOXYPROPYL)-WYOSINE BIOSYNTHESIS\%HUMANCYC\%PWY-7286") 
}

