\name{cnv_GR}
\alias{cnv_GR}
\docType{data}
\title{
CNV locations for breast cancer (subset)
}
\description{
Subset of CNV locations for TCGA breast tumour. Each range is 
 associated with a patient (ID)
}
\usage{data(cnv_GR)}
\source{
The Cancer Genome Atlas. (2012). Nature 490:61-70.
}
\references{
The Cancer Genome Atlas. (2012). Nature 490:61-70.
}
\examples{
data(cnv_GR)
head(cnv_GR)
}
\keyword{datasets}
