\name{cnv_netScores}
\alias{cnv_netScores}
\docType{data}
\title{
List of pathway-level feature selection scores
}
\description{
List of pathway-level feature selection scores
}
\usage{data(cnv_netScores)}
\examples{
data(cnv_netScores)
summary(cnv_netScores)
head(cnv_netScores[[1]])
}
\keyword{datasets}
