% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileFeatures.R
\name{compileFeatures}
\alias{compileFeatures}
\title{Create GeneMANIA database}
\usage{
compileFeatures(netDir, patientID, outDir, simMetric = "pearson",
  netSfx = "_cont.txt$", verbose = TRUE, numCores = 1L,
  P2N_threshType = "off", P2N_maxMissing = 100, JavaMemory = 4L, ...)
}
\arguments{
\item{netDir}{(char) path to dir with input networks/profiles. All
networks in this directory will be added to the GM database. Note:
This needs to be an absolute path, not relative.}

\item{patientID}{(char) vector of patient IDs.}

\item{outDir}{(char) path to dir in which GeneMANIA database is created. 
The database will be under \code{outDir/dataset}.}

\item{simMetric}{(char) similarity measure to use in converting 
profiles to interaction networks.}

\item{netSfx}{(char) pattern for finding network files in \code{netDir}.}

\item{verbose}{(logical) print messages}

\item{numCores}{(integer) num cores for parallel processing}

\item{P2N_threshType}{(char) Most users shouldn't have to change this.
ProfileToNetworkDriver's threshold option. One of "off|auto".}

\item{P2N_maxMissing}{(integer 5-100)}

\item{JavaMemory}{(integer) Memory for GeneMANIA (in Gb)}

\item{...}{params for \code{writeQueryBatchFile()}}
}
\value{
(list). "dbDir": path to GeneMANIA database 
	"netDir": path to directory with interaction networks. If profiles
are provided, this points to the INTERACTIONS/ subdirectory within 
the text-based GeneMANIA generic database
If the DB creation process results in an erorr, these values return 
NA
}
\description{
Create GeneMANIA database
}
\details{
Creates a generic_db for use with GeneMania QueryRunner.
The database is in tab-delimited format, and indexes are built using Apache lucene.
NOTE: This pipeline expects input in the form of interaction networks
and not profiles.
Profile tables have patient-by-datapoint format (e.g. patient-by-genotype)
Interaction networks have pairwise similarity measures:
<PatientA>	<PatientB>	<similarity>
Documentation: https://github.com/GeneMANIA/pipeline/wiki/GenericDb
}
\examples{
data(TCGA_mini,pathwayList);
# note: the paths in the calls below need to be absolute. If you 
# do not have write access to /tmp, change to a different directory.
n <- makePSN_NamedMatrix(xpr,rownames(xpr),pathwayList,"/tmp/nets/",
	writeProfiles=TRUE); 
db <- compileFeatures("/tmp/nets/",pheno$ID,"/tmp")
}
