% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileFeatureScores.R
\name{compileFeatureScores}
\alias{compileFeatureScores}
\title{Tally the score of networks through cross-validation}
\usage{
compileFeatureScores(fList, filter_WtSum = 100, verbose = FALSE)
}
\arguments{
\item{fList}{(char) Vector of paths to GeneMANIA NRANK files}

\item{filter_WtSum}{(numeric between 5-100) Limit to top-ranked 
networks such that cumulative weight is less than this parameter. 
e.g. If filter_WtSum=20, first order networks by decreasing weight; 
then keep those whose cumulative weight <= 20.}

\item{verbose}{(logical) print messages}
}
\value{
(data.frame) Feature name and score; includes features that occur
at least once in \code{fList}.
}
\description{
Tally the score of networks through cross-validation
}
\examples{
netDir <- system.file("extdata","GM_NRANK",package="netDx")
netFiles <- sprintf('\%s/\%s', netDir,dir(netDir,pattern='NRANK$'))
pTally <- compileFeatureScores(netFiles,verbose=TRUE)
print(head(pTally))
}
