% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooMutationPropagation.R
\name{complete_m}
\alias{complete_m}
\title{Remove genes in the dataset which don't appear in the network}
\usage{
complete_m(m_profiles, net)
}
\arguments{
\item{m_profiles}{(data.frame) sparse matrix of patient profiles. Rownames
are unique genes. Colnames are unique patients. A cell is a numeric value.}

\item{net}{(data.frame) adjancy matrix format of a network}
}
\value{
(data.frame) sparse matrix of patient profiles with a number of
  rows/genes lower or equal to its starting value
}
\description{
Remove genes in the dataset which don't appear in the network
}
\details{
This function removes the genes which appear in the patient's profiles and
  that are not represented as nodes in the network selected for the
  propagation. In fact, the network-based propagation computes a score for
  each gene/node depending on how much is connected to all the other ones. In
  case an expressed or mutated gene in the sparse matrix is not represented,
  it cannot get the new score of propagation.
}
