\name{confmat}
\alias{confmat}
\docType{data}
\title{
	Confusion matrix example
}
\description{
Sample table of True/False Positives and Negatives for various feature 
selection cutoffs
tp: true positive rate, 
fp: false positive rate, 
tn: true negative rate, 
fn: false negative rate
}
\usage{data(confmat)}
\examples{
data(confmat)
head(confmat)
}
\keyword{datasets}
