% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{make confusion matrix}
\usage{
confusionMatrix(model)
}
\arguments{
\item{(list)}{output of buildPredictor()}
}
\value{
(list) confusion matrix for all train/test splits and final averaged matrix
}
\description{
make confusion matrix
}
