% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runProfileToNetworks.R
\name{convertProfileToNetworks}
\alias{convertProfileToNetworks}
\title{Convert profiles to interaction networks before integration}
\usage{
convertProfileToNetworks(netDir, outDir = tempdir(),
  simMetric = "pearson", numCores = 1L, JavaMemory = 4L,
  GM_jar = NULL, P2N_threshType = "off", P2N_maxMissing = 100,
  netSfx = "txt$", debugMode = FALSE)
}
\arguments{
\item{netDir}{(char) directory with .profile files}

\item{outDir}{(char) path to directory where interaction networks are to be printed}

\item{simMetric}{(char) similarity measure to use in converting 
profiles to interaction networks.}

\item{numCores}{(integer) number of cores for parallel processing}

\item{JavaMemory}{(integer) Memory for GeneMANIA (in Gb)}

\item{GM_jar}{(char) path to GeneMANIA jar file}

\item{P2N_threshType}{(char) Most users shouldn't have to change this.
ProfileToNetworkDriver's threshold option. One of 'off|auto'. 
unit testing}

\item{P2N_maxMissing}{(integer 5-100)}

\item{netSfx}{(char) pattern for finding network files in \code{netDir}.}

\item{debugMode}{(logical) if TRUE runs profile generation in serial 
rather than parallel, allowing debugging}
}
\value{
No value. Side effect of creating interaction networks in outDir.
}
\description{
Convert profiles to interaction networks before integration
}
\details{
In preparation for network integration. When using GeneMANIA's
built-in functionality to create PSN using ProfileToNetworkDriver, this
step needs to run to process profiles to networks. These are currently used
for Pearson correlation-based networks and those using mutual information.
}
