% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countIntType.R
\name{countIntType}
\alias{countIntType}
\title{Counts the number of (+,+) and (+,-) interactions in a single network}
\usage{
countIntType(inFile, plusID, minusID)
}
\arguments{
\item{inFile}{(char) path to interaction networks}

\item{plusID}{(char) vector of + nodes}

\item{minusID}{(char) vector of - nodes}
}
\description{
Counts the number of (+,+) and (+,-) interactions in a single network
}

