% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{createInputForFeatureNetworkView}
\alias{createInputForFeatureNetworkView}
\title{Wrapper to create input files for Enrichment Map}
\usage{
createInputForFeatureNetworkView(
  model,
  results,
  pathwayList,
  EMapMinScore = 0L,
  EMapMaxScore = 1L,
  EMapPctPass = 0.5,
  outDir
)
}
\arguments{
\item{model}{(list) Output of training model, generated by running buildPredictor()}

\item{results}{(list) Model results. output of getResults()}

\item{pathwayList}{(list) output of readPathwayFile() used to make pathway-level feat ures for predictor}

\item{EMapMinScore}{(integer) minimum score for Enrichment Map}

\item{EMapMaxScore}{(integer) maximum score for Enrichment Map}

\item{EMapPctPass}{(numeric between 0 and 1) percent of splits for which feature must have score in range
[EMapMinScore,EMapMaxScore] to be included for EnrichmentMap visualization}

\item{outDir}{(char) directory where files should be written}
}
\value{
(list) 1) GMTfiles (char): GMT files used to create EnrichmentMap in Cytoscape.
2) NodeStyles (char): .txt files used to assign node attributes in Cytoscape. Importantly, 
attributes include node fill, which indicates the highest consistent score for a given 
feature.
}
\description{
Wrapper to create input files for Enrichment Map
}
\details{
Creates the input to visualize selected features and their relationships
as a network in Cytoscape. The type of visualization is called an Enrichment Map.
An Enrichment Map is a network-based visualization of top-scoring pathway features
and themes. It is generated in Cytoscape. This script generates the input files needed
for Cytoscape to create an Enrichment Map visualization.
}
