% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{createMultiAssayExperiment}
\alias{createMultiAssayExperiment}
\title{creates a MultiAssayExperiment object from user-provided files}
\usage{
createMultiAssayExperiment(dataTables, pheno)
}
\arguments{
\item{dataTables}{(list of matrices/data.frames) one key per data type, 
with table in corresponding value. Keys are user-defined names for the data layers}

\item{pheno}{(data.frame) sample metadata. Rows are patients and columns are variables.
* Patient IDs must be in rownames}
}
\description{
creates a MultiAssayExperiment object from user-provided files
}
