% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{createNetFuncFromSimList}
\alias{createNetFuncFromSimList}
\title{create nets from provided similarities}
\usage{
createNetFuncFromSimList(
  dataList,
  groupList,
  netDir,
  sims,
  verbose = TRUE,
  ...
)
}
\description{
create nets from provided similarities
}
