% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPSN_MultiData.R
\name{createPSN_MultiData}
\alias{createPSN_MultiData}
\title{Wrapper to create custom input features (patient similarity networks)}
\usage{
createPSN_MultiData(dataList, groupList, netDir, filterSet = NULL,
  verbose = TRUE, customFunc, ...)
}
\arguments{
\item{dataList}{(list) key is datatype (e.g. clinical, rna, etc.,), value is
 table or RangedData)
Note that unit names should be rownames of the data structure.
e.g If dataList$rna contains genes, rownames(dataList) = gene names}

\item{groupList}{(list) key is datatype; value is a list of unit groupings
for that datatype. e.g. If rna data will be grouped by pathways, then 
groupList$rna would have pathway names as keys, and member genes as units.
Each entry will be converted into a PSN.}

\item{netDir}{(char) path to directory where networks will be stored}

\item{filterSet}{(char) vector of networks to include}

\item{verbose}{(logical) print messages}

\item{customFunc}{(function) custom user-function to create PSN. 
Must take dataList,groupList,netDir as parameters. Must
check if a given groupList is empty (no networks to create) before 
the makePSN call for it. This is to avoid trying to make nets for datatypes
that did not pass feature selection}

\item{...}{other parameters to makePSN_NamedMatrix() or makePSN_RangedSets()}
}
\value{
(char) vector of network names. Side effect of creating the nets
}
\description{
Wrapper to create custom input features (patient similarity networks)
}
\examples{
createPSN_MultiData(dataList=dats_train,groupList=groupList,
netDir=netDir,customFunc=makeNetFunc,numCores=numCores,
verbose=verbose_default)
}
