% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileCache.R
\name{.get_cache}
\alias{.get_cache}
\title{wrapper function for getting BiocFileCache associated with netDx package}
\usage{
.get_cache()
}
\value{
BiocFileCache object associated with netDx
}
\description{
wrapper function for getting BiocFileCache associated with netDx package
}
