% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileCache.R
\name{fetchPathwayDefinitions}
\alias{fetchPathwayDefinitions}
\title{fetch pathway definitions from downloads.baderlab.org}
\usage{
fetchPathwayDefinitions(month = NULL, year = NULL, verbose = FALSE)
}
\arguments{
\item{month}{(char) month of pathway definition file. Must be
textual name (e.g. "January","April"). If NULL, gets the latest
release}

\item{year}{(numeric) year of pathway definition file. Must be in
yyyy format (e.g. 2018).}

\item{verbose}{(logical) print messages}
}
\value{
(char) Path to local cached copy of GMT file
or initial download is required
}
\description{
fetch pathway definitions from downloads.baderlab.org
}
\details{
Fetches genesets compiled from multiple curated pathway
databases. Downloaded from: http://download.baderlab.org/EM_Genesets/
The file contains pathways from HumanCyc, NetPath, Reactome, NCI
Curated Pathways and mSigDB.
For details see Merico D, Isserlin R, Stueker O, Emili A and GD Bader.
(2010). PLoS One. 5(11):e13984.
}
\examples{
fetchPathwayDefinitions()
fetchPathwayDefinitions("January","2018")
fetchPathwayDefinitions()
}
