\name{genes}
\alias{genes}
\docType{data}
\title{
	Table of gene definitions (small subsample of human genes)
}
\description{
data.frame object with columns of (gene) RefSeq ID (name), chromosome (chrom), strand, transcription start site (txStart), transcription end site (txEnd), and gene symbol (name2)
}
\usage{data(genes)}
\examples{
data(genes)
head(genes)
}
\keyword{datasets}
