% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAUC_OLnets.R
\name{getAUC_OLnets}
\alias{getAUC_OLnets}
\title{get AUC for shared overlap nets}
\usage{
getAUC_OLnets(inFile)
}
\arguments{
\item{inFile}{(char) RR_changeNetSum_stats_den*Nets.txt. Contains a table
with pred/other OL for each score cutoff. Output of RR_featureTally()}
}
\value{
(list) performance measures: 
1) stats: (data.frame) ppv,f1,rec for each cutoff
2) auc: (numeric) AUC-ROC
3) prauc: (numeric) AUC-PR
}
\description{
get AUC for shared overlap nets
}
\details{
this predictor doesn't use GM to rank patient similarity. 
Rather a "called +" is a patient with CNVs that overlap feature-selected
nets
}
