% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorrType.R
\name{getCorrType}
\alias{getCorrType}
\title{Counts the relative correlation of (+,+) and (+,-)(-,-) interactions}
\usage{
getCorrType(inFile, plusID, minusID)
}
\arguments{
\item{inFile}{(character): path to interaction networks}

\item{plusID}{(character) vector of + nodes}

\item{minusID}{(character) vector of - nodes}
}
\value{
(numeric) mean edge weight for (+,+) and other edges
}
\description{
Counts the relative correlation of (+,+) and (+,-)(-,-) interactions
}
