% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnr.R
\name{getEnr}
\alias{getEnr}
\title{Get ENR for all networks in a specified directory}
\usage{
getEnr(netDir, pheno_DF, predClass, netGrep = "_cont.txt$",
  enrType = "binary", ...)
}
\arguments{
\item{netDir}{(char) directory containing interaction networks}

\item{pheno_DF}{(data.frame) table with patient ID and status.
Must contain columns for Patient ID (named "ID") and class
(named "STATUS"). Status should be a char; value of predictor class 
should be specified in \code{predClass} param; 
all other values are considered non-predictor class
Rows with duplicate IDs will be excluded.}

\item{predClass}{(char) value for patients in predictor class}

\item{netGrep}{(char) pattern for grep-ing network text files, used in
dir(pattern=..) argument}

\item{enrType}{(char) how enrichment should be computed. Options are:
1) binary: Skew of number of (+,+) interactions relative to other
interactions. Used when all edges in network are set to 1 (e.g. 
shared CNV overlap)
2) corr: 0.5*((mean weight of (+,+) edges)-(mean weight of other edges))}

\item{...}{arguments for \code{countIntType_batch}}
}
\value{
(list):
1) plusID (char) vector of + nodes
2) minusID (char) vector of - nodes
3) orig_rat (numeric) \code{ENR} for data networks
4) fList (char) set of networks processed
5) orig (data.frame) output of \code{countIntType_batch} for input
networks
}
\description{
Get ENR for all networks in a specified directory
}
\details{
For each network, compute the number of (+,+) and other 
{(+,-),(-,+),(-,-)} interactions. 
 From this compute network ENR.
The measure of (+,+)-enrichment is defined as: 
ENR(network N) = ((num (+,+) edges) - (num other edges))/(num edges).
A network with only (+,+) interactions has an ENR=1 ; a network with
no (+,+) interactions has an ENR=-1; a network with a balance of the two
has ENR=0.
}
\examples{
data(npheno)
netDir <- sprintf("\%s/extdata/example_nets",path.package("netDx"))
x <- getEnr(netDir, npheno, "case",netGrep=".txt$")
print(x$orig_rat) 

}
