% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileCache.R
\name{getExamplePathways}
\alias{getExamplePathways}
\title{download pathway gmt file used for several examples}
\usage{
getExamplePathways(verbose = FALSE)
}
\arguments{
\item{verbose}{(logical) print messages}
}
\value{
(char) Path to local cached copy of GMT file
or initial download is required
}
\description{
download pathway gmt file used for several examples
}
\examples{
getExamplePathways()
}
