% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatureScores.R
\name{getFeatureScores}
\alias{getFeatureScores}
\title{Compile network scores into a matrix}
\usage{
getFeatureScores(inDir, predClasses)
}
\arguments{
\item{inDir}{(char/list) directory containing directories with all split info
or list of all CV score files.
if inDir is a single directory then the expected format for CV score files is
<inDir>/rngX/predClassX/GM_results/predClassX_pathway_CV_score.txt"
if inDir is a list, it should have one key per class. The value should be the
corresponding set of filenames for pathway_CV_score.txt}

\item{predClasses}{(char) possible STATUS for patients}
}
\value{
(list) one key per patient class. Value is matrix of network
scores across all train/test splits. Each score is the output of
the inner fold of CV.
}
\description{
Compile network scores into a matrix
}
\details{
Given network scores over a set of trian/test splits, compiles these
into a matrix for downstream analysis. See the section on "Output Files"
}
\examples{
inDir <- sprintf("\%s/inst/extdata/KIRC_example_data", 
		path.package("netDx_software_update"))
all_rngs <- list.dirs(inDir, recursive = FALSE)
all_classone_files <- unlist(lapply(all_rngs, function(x) 
		paste(x, "SURVIVEYES/GM_results/SURVIVEYES_pathway_CV_score.txt", 
	sep = "/")))
 all_classtwo_files <- unlist(lapply(all_rngs, function(x) 
		paste(x, "SURVIVENO/GM_results/SURVIVENO_pathway_CV_score.txt", 
		sep = "/")))
 input_list <- list("SURVIVEYES" = all_classone_files,
                     "SURVIVENO" = all_classtwo_files)
net_scores <- genNetScores(inDir, c("SURVIVEYES","SURVIVENO"))
}
