% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFileSep}
\alias{getFileSep}
\title{platform-specific file separator}
\usage{
getFileSep()
}
\value{
(char) "\\" if Windows, else "/"
}
\description{
Returns OS-specific file separator
}
\examples{
getFileSep()
}
