% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileCache.R
\name{getGMjar_path}
\alias{getGMjar_path}
\title{download and update GeneMANIA jar file}
\usage{
getGMjar_path(verbose = FALSE)
}
\arguments{
\item{verbose}{(logical) print messages}
}
\value{
(char) Path to local cached copy of GeneMANIA jar file..
or initial download is required
}
\description{
download and update GeneMANIA jar file
}
\examples{
getGMjar_path()
}
