% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetConsensus.R
\name{getNetConsensus}
\alias{getNetConsensus}
\title{compile net score across a set of predictor results}
\usage{
getNetConsensus(scorelist)
}
\arguments{
\item{scorelist}{(list) key is dataset name, value is a data.frame 
containing PATHWAY_NAME and SCORE. This is the output of
 compileFeatureScores()}
}
\value{
(data.frame) Rownames are union of all nets in the input list.
Columns show net scores for each key of the input list. Where a 
net is not found in a given list, it is assigned the value of NA
}
\description{
compile net score across a set of predictor results
}
\details{
used to compare how individual nets score for different
predictor configurations
}
\examples{
pathways <- paste("PATHWAY_",1:100,sep="")
highrisk <- list()
for (k in 1:10) { 
	highrisk[[k]] <- data.frame(PATHWAY_NAME=pathways, 
	SCORE=runif(length(pathways),min=0,max=10),
			stringsAsFactors=FALSE);
}
names(highrisk) <- sprintf("Split\%i",1:length(highrisk))
x <- getNetConsensus(highrisk)
}
