% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOR.R
\name{getOR}
\alias{getOR}
\title{Get relative proportion of patient classes that contribute to a set of
networks}
\usage{
getOR(pNetworks, pheno_DF, predClass, netList, verbose = TRUE)
}
\arguments{
\item{pNetworks}{(matrix) rows are patients, columns are network file
filenames. a[i,j] = 1 if patient i has a structural variant in network
j; else a[i,j] = 0}

\item{pheno_DF}{(data.frame) Column "ID" has unique patient identifiers;
column "STATUS" has patient class}

\item{predClass}{(char) Class for which predictor is being built}

\item{netList}{(char) vector of networks of interest (e.g. those 
passing feature selection)}

\item{verbose}{(logical) print messages}
}
\value{
List. 1) stats: statistics on group overlap with netList,
2) relEnr: relative enrichment of \code{predClass} over other
}
\description{
Get relative proportion of patient classes that contribute to a set of
networks
}
\details{
Feature selected networks should have the property of being
enriched in the class of interest; e.g. be enriched in 'case' relative
to 'control'. When given a list of networks N, this method computes the
number and proportion of patients that overlap N. A high relative 
fraction of the predicted class indicates successful feature selection.
To create a ROC or precision-recall curve, several calls can be made
to this function, one per cutoff.
}
\examples{
data(npheno)
netDir <- sprintf("\%s/extdata/example_nets",path.package("netDx"))
x <- countPatientsInNet(netDir,dir(netDir,pattern="txt$"), npheno[,1])
getOR(x,npheno,"case",colnames(x)[1]) # should give large RelEnr
getOR(x,npheno,"case",colnames(x)[2]) # should give RelEnr of 0
getOR(x,npheno,"case",colnames(x)[3]) # should give RelEnr of 1
}

