% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPatientPredictions.R
\name{getPatientPredictions}
\alias{getPatientPredictions}
\title{Calculates patient-level classification accuracy across train/test splits}
\usage{
getPatientPredictions(predFiles, pheno)
}
\arguments{
\item{predFiles}{(char) vector of paths to all test predictions
(e.g. 100 files for a 100 train/test split design).
Alternately, the user can also  provide a single directory name, and allow
the script to retrieve prediction files.
Format is "rootDir/rngX/predictionResults.txt"}

\item{pheno}{(data.frame) ID=patient ID, STATUS=ground truth (known class
label). This table is required to get the master list of all patients, as
not every patient is classified in every split.}
}
\value{
(list) of length 2.
1) (data.frame) rows are patients, (length(predFiles)+2) columns.
Columns 1:length(predFiles): Predicted labels for a given split (NA if
patient was training sample for the split).
Column (length(predFiles)+1):
split, value is NA. Columns are : ID, REAL_STATUS, predStatus1,...
predStatusN.
Side effect of plotting a dot plot of % accuracy. Each dot is a patient, and
the value is "% splits for which patient was classified correctly".
}
\description{
Calculates patient-level classification accuracy across train/test splits
}
\details{
Takes all the predictions across the different train/test splits,
and for each patient, generates a score indicating how many times they were
classified by netDx as belonging to each of the classes. The result is that
we get a measure of individual classification accuracy across the different
train/test splits.
}
\examples{
inDir <- sprintf("\%s/extdata/KIRC_output",
   path.package("netDx.examples"))
phenoFile <- sprintf("\%s/extdata/KIRC_pheno.rda",path.package("netDx.examples"))
load(phenoFile)
all_rngs <- list.dirs(inDir, recursive = FALSE)
all_pred_files <- unlist(lapply(all_rngs, function(x) {
    paste(x, "predictionResults.txt", sep = "/")}))
pred_mat <- getPatientPredictions(all_pred_files, pheno)
}
