% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPerf_simple.R
\name{getPerformance}
\alias{getPerformance}
\title{performance metrics for model}
\usage{
getPerformance(res, predClasses)
}
\arguments{
\item{res}{(resa.frame) result from predicting labels on held-out test set. output of predict() function. 
columns include ID, STATUS (ground truth) and PRED_CLASS (predicted label)}
}
\value{
(list)
1) rocCurve: ROCR performance object for ROC curve
2) prCurve: ROCR performance object for PR curve
3) auroc: Area under ROC curve
4) aupr: Area under PR curve
5) accuracy: Accuracy
}
\description{
performance metrics for model
}
