% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getPSN}
\alias{getPSN}
\title{get the integrated patient similarity network made of selected features}
\usage{
getPSN(
  dat,
  groupList,
  makeNets,
  selectedFeatures,
  plotCytoscape = FALSE,
  aggFun = "MEAN",
  prune_pctX = 0.3,
  prune_useTop = TRUE,
  numCores = 1L,
  calcShortestPath = FALSE
)
}
\arguments{
\item{dat}{(MultiAssayExperiment) input data}

\item{aggFun}{- see plotIntegratedPatientNetwork()}

\item{prune_pctX}{- see plotIntegratedPatientNetwork()}

\item{prune_useTop}{- see plotIntegratedPatientNetwork()}

\item{numCores}{- see plotIntegratedPatientNetwork()}

\item{calcShortestPath}{- see plotIntegratedPatientNetwork()}

\item{res}{(list) output of buildPredictor() function}
}
\value{
(list) output of plotIntegratedPatientNetwork()
}
\description{
get the integrated patient similarity network made of selected features
}
