% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegionOL.R
\name{getRegionOL}
\alias{getRegionOL}
\title{Returns overlapping named ranges for input ranges}
\usage{
getRegionOL(gr, rngList)
}
\arguments{
\item{gr}{(GRanges) query ranges}

\item{rngList}{(list) keys are names, and values are GRanges, each range
of which has a name (in "name" column). Note: It is faster to provide
a list of length 1 ; if the list is long, combining into a single GRanges
object could prove slow.}
}
\value{
(GRanges) query ranges with the added column "LOCUS_NAMES". 
Where a range overlaps with multiple loci, the names are reported as a 
comma-separated vector
}
\description{
Returns overlapping named ranges for input ranges
}
\details{
Given a set of query GRanges, and a subject list-of-GRanges,
updates the query with a column 'LOCUS_NAMES' containing the names of
ranges overlapped by the query. One application is to map structural
variants, such as CNVs, to genes in pathways of interest. In this 
scenario \code{gr} would contain the patient CNVs, and \code{rngList}
would be a list of GenomicRanges objects, one per cellular pathway.
}
\examples{
data(xpr,pheno,cnv_GR,pathway_GR)
x <- getRegionOL(cnv_GR,pathway_GR)
}
