% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFun.R
\name{getSimilarity}
\alias{getSimilarity}
\title{Measures of patient similarity}
\usage{
getSimilarity(x, type = "pearson", customFunc, ...)
}
\arguments{
\item{x}{(matrix) matrix for which pairwise patient similarity is to be
computed. Expects one column per patient, and one measurement per row.}

\item{type}{(character) name of similarity measure. Currently supports 
Pearson correlation ("pearson") or a custom measure ("custom")}

\item{customFunc}{(function) custom similarity function. Only used when 
\code{type="custom"}. The function takes \code{x} as first argument and 
can take additional argument. It should return a symmetric matrix of 
pairwise patient similarities.}
}
\value{
symmetric matrix of size N, where N is number of samples
}
\description{
Measures of patient similarity
}
\examples{
data(TCGA_mini) 
x <- getSimilarity(xpr) # similarity by Pearson corr
mySim <- function(x) cor(x,method="kendall")
x <- getSimilarity(xpr,customFunc=mySim) # custom similarity
}

