% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_getQueryROC.R
\name{GM_getQueryROC}
\alias{GM_getQueryROC}
\title{Process GM PRANK files to get the ROC curve for the query}
\usage{
GM_getQueryROC(pFile, pheno_DF, predClass, plotIt = FALSE,
  verbose = FALSE)
}
\arguments{
\item{pFile}{(char) path to PRANK file}

\item{pheno_DF}{(data.frame) patient IDs ("ID") and label("STATUS")}

\item{predClass}{(character) class label for which predictor is built}

\item{plotIt}{(logical) if TRUE plots ROC curve}

\item{verbose}{(logical) print messages}
}
\value{
(list) 
1) predLbl: GeneMANIA scores (predicted labels). Higher score for
higher ranked patient. 
2) realLbl: binary value indicating if patient label matches predictor
label (real labels)
3) fullmat: pheno_DF merged with GM scores ("GM_score") and real label
("isPredClass")
4) roc: output of ROCRs performance(,"tpr","fpr") - ROC curve
5) auc: output of ROCRs auc() 
6) precall: output of ROCRs performance(, "prec","rec")
7) f: output of ROCRs performance(,"f")
If < 2 patients in PRANK file, roc,auc, precall, f are all returned as
NA.
}
\description{
Process GM PRANK files to get the ROC curve for the query
}
\examples{
data(TCGA_mini)
prankFile <- sprintf("\%s/extdata/GM_PRANK.txt", 
	 path.package("netDx"))
x <- GM_getQueryROC(prankFile, pheno, "LumA")
}
