% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_netIDtoNames.R
\name{GM_netIDtoNames}
\alias{GM_netIDtoNames}
\title{resolve patient number in GeneMANIA interaction net to names}
\usage{
GM_netIDtoNames(netFile, geneFile)
}
\arguments{
\item{netFile}{(char) path to GeneMANIA interaction net. SOURCE TARGET
and WEIGHT. Source and target should have GM internal ID}

\item{geneFile}{(char) path to GeneMANIA GENES.txt file which resolves
internal (GM-assigned) patient ID to names}
}
\value{
(data.frame) network with SOURCE TARGET WEIGHT, but with 
original patient ID
}
\description{
resolve patient number in GeneMANIA interaction net to names
}

