% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_networkTally.R
\name{GM_networkTally}
\alias{GM_networkTally}
\title{Tally the score of networks through cross-validation}
\usage{
GM_networkTally(fList, filter_WtSum = 100, verbose = FALSE,
  useNewGM = FALSE)
}
\arguments{
\item{fList}{(char) Vector of paths to GeneMANIA NRANK files}

\item{filter_WtSum}{(numeric between 5-100) Limit to top-ranked 
networks such that cumulative weight is less than this parameter. 
e.g. If filter_WtSum=20, first order networks by decreasing weight; 
then keep those whose cumulative weight <= 20.}

\item{verbose}{(logical) print messages}

\item{useNewGM}{(logical) use new GeneMania Version for processing - 
slightly different output format}
}
\value{
(named integer) Vector of scores for networks that occur at 
least once in \code{fList}.
}
\description{
Tally the score of networks through cross-validation
}
\examples{
netDir <- sprintf("\%s/extdata/GM_NRANK",path.package("netDx"))
netFiles <- sprintf("\%s/\%s", netDir,dir(netDir,pattern="NRANK$"))
pTally <- GM_networkTally(netFiles)
print(head(pTally))
}
