% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_OneVAll_getClass.R
\name{GM_OneVAll_getClass}
\alias{GM_OneVAll_getClass}
\title{assign patient class when ranked by multiple GM predictors}
\usage{
GM_OneVAll_getClass(resSet)
}
\arguments{
\item{resSet}{(list) output of GM_getQueryROC, each key for a different
predictor. names(resSet) contain predictor label}
}
\value{
data.frame: ID, GM_score, PRED_CLASS
}
\description{
assign patient class when ranked by multiple GM predictors
}
\examples{
data(predRes); predClass <- GM_OneVAll_getClass(predRes)
}
