% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_parseReport.R
\name{GM_parseReport}
\alias{GM_parseReport}
\title{parses GeneMANIA query report into the PRANK and NRANK tables}
\usage{
GM_parseReport(resFile)
}
\arguments{
\item{resFile}{(char) path to GeneMANIA query results}
}
\value{
No value. Side effect of writing PRANK and NRANK files to the 
same directory as resFile. Suffixes filename with PRANK and NRANK 
respectively
}
\description{
parses GeneMANIA query report into the PRANK and NRANK tables
}

