% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_runCV_featureSet.R
\name{GM_runCV_featureSet}
\alias{GM_runCV_featureSet}
\title{Run GeneMANIA cross-validation with a provided subset of networks}
\usage{
GM_runCV_featureSet(trainID_pred, outDir, GM_db, numTrainSamps = NULL,
  incNets = "all", orgName = "predictor", fileSfx = "CV",
  verbose = FALSE, numCores = 2L, GMmemory = 6L, seed_CVqueries = 42L,
  ...)
}
\arguments{
\item{trainID_pred}{(char) vector with universe of predictor class
patients (ie all that can possibly be included in the query file}

\item{outDir}{(char) directory to store query file and GM results}

\item{GM_db}{(char) path to GeneMANIA generic database with
training population}

\item{numTrainSamps}{(integer) number of training samples in total
leave blank to use 5 training samples in order to save memory}

\item{incNets}{(char) vector of networks to include in this analysis
(features/pathway names). Useful for subset-based feature selection}

\item{orgName}{(char) organism name for GeneMANIA generic database.
The default value will likely never need to be changed.}

\item{fileSfx}{(char) file suffix}

\item{verbose}{(logical) print messages}

\item{numCores}{(logical) num parallel threads for cross-validation}

\item{GMmemory}{(integer) memory for GeneMANIA run, in Gb.}

\item{seed_CVqueries}{(integer) RNG seed for inner cross validation loop.
Makes deterministic samples held-out for each GeneMANIA query (see
makeCVqueries())}

\item{...}{args for \code{makeCVqueries()}}
}
\description{
Run GeneMANIA cross-validation with a provided subset of networks
}
\details{
Creates query files, runs GM for 10-fold cross validation.
}
\examples{
data(MB_pheno)
GM_db <- sprintf("\%s/extdata/GM_db",path.package("netDx"))
GM_runCV_featureSet(MB.pheno$ID[which(MB.pheno$STATUS\%in\% "WNT")],
"~/tmp",GM_db,103L)
}
