% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_runCV_nested.R
\name{GM_runCV_nested}
\alias{GM_runCV_nested}
\title{Nested cross validation for GM feature selection}
\usage{
GM_runCV_nested(nOuterLoop = 10L, setOloopSeed = 42L, outDir, pheno_DF,
  predictClass, ...)
}
\arguments{
\item{nOuterLoop}{(integer) number of outer loops for cross validation}

\item{setOloopSeed}{(integer) seed for random number generator that
in turn sets the seed for inner CV loop. The RNG samples numbers from
a vector 1:1000. Note: regardless, this value overrides the 
\code{setSeed} param of \code{makeCVqueries}.}

\item{outDir}{(char) directory to store results of GeneMANIA queries}

\item{pheno_DF}{(data.frame) patient ID and STATUS}

\item{predictClass}{(char) class (of pheno_DF$STATUS) being predicted}

\item{...}{parameters for runFeatureSelection()}
}
\value{
(char) vector of paths to directories where result for each
outer loop is stored.
}
\description{
Nested cross validation for GM feature selection
}
\details{
Runs two-tier cross validation by calling
\code{runFeatureSelection()} for the inner loop. Computes patient
ranking and network tally
}
\examples{
data(MB_pheno)
dbPath <- sprintf("\%s/extdata/dbPath",path.package("netDx"))
GM_runCV_nested(outDir=".",pheno_DF=MB.pheno, predictClass="WNT",
	trainID_pred=MB.pheno$ID[which(MB.pheno$STATUS\%in\% "WNT")],
	numTrainSamps=103L,dbPath=dbPath)
}
