% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_writeBatchFile.R
\name{GM_writeBatchFile}
\alias{GM_writeBatchFile}
\title{Write batch.txt file required to create GeneMANIA database}
\usage{
GM_writeBatchFile(netDir, netList, outDir, idFile, orgName = "predictor",
  orgDesc = "my_predictor", orgAlias = "my_predictor", taxID = 1339)
}
\arguments{
\item{netDir}{(char) path to dir with networks}

\item{netList}{(char) vector of network names}

\item{outDir}{(char) directory to write batch file}

\item{idFile}{(char) path to file with patient IDs}

\item{orgName}{(char) organism name. Don't change the default unless
you know what you are doing.}

\item{orgDesc}{(char) organism description. Similar to \code{orgName},
don't change the default}

\item{orgAlias}{(char) organism alias. Similar to \code{orgName}, don't
change the default.}

\item{taxID}{(integer) taxonomyID required for GeneMANIA . Similar to 
\code{orgName}, don't change the default.}
}
\value{
No value. Side effect of writing batch file to 
\code{<outDir>/batch.txt}.
}
\description{
Write batch.txt file required to create GeneMANIA database
}
\details{
Currently creates a single entry for network group. Future versions of this
package may want to
}
\examples{
data(npheno)
netDir <- sprintf("\%s/extdata/example_nets",path.package("netDx"))
netList <- dir(netDir,pattern="txt$")
GM_writeBatchFile(netDir,netList, "~/tmp", npheno$ID)
}
