% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_writeQueryFile.R
\name{GM_writeQueryFile}
\alias{GM_writeQueryFile}
\title{Wrapper to write GeneMANIA query file}
\usage{
GM_writeQueryFile(qSamps, incNets, numReturn = 1L, outFile,
  orgName = "predictor")
}
\arguments{
\item{qSamps}{(char) vector of patient IDs in query}

\item{incNets}{(char) vector of networks to include in this analysis 
(features/pathway names). Useful for subset-based feature selection}

\item{numReturn}{(integer) number of patients to return in ranking file}

\item{outFile}{(char) path to output file}

\item{orgName}{(char) organism name}
}
\value{
No value. Side effect of writing the query file to
\code{outFile}
}
\description{
Wrapper to write GeneMANIA query file
}

