\name{KIRC_dat}
\alias{KIRC_dat}
\docType{data}
\title{
Clinical and gene expression data for kidney cancer survival.
}
\description{
List with one entry for clinical data ("clinical") and one for gene expression ("rna"). Each entry contains a data.frame with patient-level measures (columns) for clinical variables or genes (rows).
}
\usage{data(KIRC_dat)}
\source{

}
\references{
Yuan Y, Van Allen EM, Omberg L, Wagle N, Amin-Mansour A, Sokolov A, Byers LA, Xu Y, Hess KR, Diao L, Han L, Huang X, Lawrence MS, Weinstein JN, Stuart JM, Mills GB, Garraway LA, Margolin AA, Getz G, Liang H (2014b) Synapse syn1710286. 
}
\examples{
data(KIRC_dat)
head(KIRC_dat)
}
\keyword{datasets}
