\name{KIRC_group}
\alias{KIRC_group}
\docType{data}
\title{
Variable groupings for kidney cancer survival data (clinical and gene expression)
}
\description{
List with one entry for clinical data ("clinical") and one for gene expression ("rna"). Each entry contains a list with keys being group names, and values containing members for the corresponding groups. Correspond to data present in KIRC_dat
}
\usage{data(KIRC_group)}
\source{

}
\references{
Yuan Y, Van Allen EM, Omberg L, Wagle N, Amin-Mansour A, Sokolov A, Byers LA, Xu Y, Hess KR, Diao L, Han L, Huang X, Lawrence MS, Weinstein JN, Stuart JM, Mills GB, Garraway LA, Margolin AA, Getz G, Liang H (2014b) Synapse syn1710286. 
}
\examples{
data(KIRC_group)
head(KIRC_group)
}
\keyword{datasets}
