\name{KIRC_pheno}
\alias{KIRC_pheno}
\docType{data}
\title{
Sample metadata table for kidney cancer example
}
\description{
data.frame with patient ID (ID), survival status (STATUS) and covariates. Data from the PanCancer Survival project. 
}
\usage{data(KIRC_pheno)}
\source{

}
\references{
Yuan Y, Van Allen EM, Omberg L, Wagle N, Amin-Mansour A, Sokolov A, Byers LA, Xu Y, Hess KR, Diao L, Han L, Huang X, Lawrence MS, Weinstein JN, Stuart JM, Mills GB, Garraway LA, Margolin AA, Getz G, Liang H (2014b) Synapse syn1710286. 
}
\examples{
data(KIRC_pheno)
head(KIRC_pheno)
}
\keyword{datasets}
