% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCVqueries.R
\name{makeCVqueries}
\alias{makeCVqueries}
\title{Randomly select patients for query in cross-validation}
\usage{
makeCVqueries(incPat, nFold = 10L, setSeed = 42L, verbose = TRUE)
}
\arguments{
\item{incPat}{(char) vector of patient IDs to be included in query}

\item{nFold}{(integer) n for n-fold cross validation}

\item{setSeed}{(integer) Set this to an integer to make results
reproducible. If NULL, seed is not explicitly set.}

\item{verbose}{(logical) print messages}
}
\value{
(list) of length \code{nFold}, containing names of patients in
query file for each fold
}
\description{
Randomly select patients for query in cross-validation
}
\examples{
data(TCGA_mini)
x <- makeCVqueries(pheno$ID)
}

