% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{makeInputForEnrichmentMap}
\alias{makeInputForEnrichmentMap}
\title{wrapper to create input files for Enrichment Map}
\usage{
makeInputForEnrichmentMap(
  model,
  results,
  pathwayList,
  EMapMinScore = 0L,
  EMapMaxSore = 1L,
  EMapPctPass = 0.5,
  outDir
)
}
\arguments{
\item{model}{(list) output of buildPredictor() function}

\item{pathwayList}{(list) output of readPathwayFile() used to make pathway-level features for predictor}

\item{EMapMinScore}{(integer) minimum score for Enrichment Map}

\item{EMapPctPass}{(numeric between 0 and 1) percent of splits for which feature must have score in range
[EMapMinScore,EMapMaxScore] to be included for EnrichmentMap visualization}

\item{featScores}{(list) keys are classes, values are matrices with feature scores for all 
train/test splits. Output of runOverallFeatureSelection()}

\item{EMapMaxScore}{(integer) maximum score for Enrichment Map}
}
\value{

}
\description{
wrapper to create input files for Enrichment Map
}
