% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePSN_RangeSets.R
\name{makePSN_RangeSets}
\alias{makePSN_RangeSets}
\title{Create patient similarity interaction networks based on range sets}
\usage{
makePSN_RangeSets(gr, rangeSet, netDir, simMetric = "coincide",
  quorum = 2L, verbose = TRUE, numCores = 1L)
}
\arguments{
\item{gr}{(GRanges) patient ranges. Metadata should contain:
ID: (char) unique patient ID
LOCUS_NAME: (comma-separated char) named ranges overlapped}

\item{rangeSet}{(list) list of GRanges, one entry per range set.
    Key is the name of the range set, and value is a GRanges object with
corresponding ranges}

\item{netDir}{(char) path to directory where networks should be written}

\item{simMetric}{(char) Similarity metric. Currently only 'coincide' 
is supported; two patients share an edge if they overlap elements in the
the same gene set. E.g. Two patients with CNVs that overlap different
genes of the same pathway would be related, but patients overlapping
genes that don't share a pathway (or, more accurately, a named-set 
grouping) would not be related. The edge weight is therefore binary.}

\item{quorum}{(integer) minimum number of patients in a network for the 
network to be constructed}

\item{verbose}{(logical) print detailed messages}

\item{numCores}{(integer) num cores for parallel processing}
}
\value{
Vector of network filenames
}
\description{
Create patient similarity interaction networks based on range sets
}
\details{
Creates patient similarity networks when data consist of 
genomic events associated with patients. Examples include CNV or 
indel data for patients. To generate networks from full matrices such
gene expression data, use \code{makePSN_NamedMatrix} instead.
Genomic ranges corresponding to events in patients (gr) should be named.
One network is created per named range set (rangeSet). Each set
reflects a group of related loci ; for example, genomic ranges associated
with genes in the same cellular pathway. 
Currently, the only similarity measure supported is binary; two patients
are related in a network N if they both overlap elements of set N.
}
\examples{
data(pathway_GR,xpr,pheno,cnv_GR)
netList <- makePSN_RangeSets(cnv_GR,pathway_GR,".")
}
