% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRandomNetworks.R
\name{makeRandomNetworks}
\alias{makeRandomNetworks}
\title{Generates random binary interaction networks}
\usage{
makeRandomNetworks(id, numNets = 100L, minNodes = 2L, maxNodes = 20L,
  sizeProb = "exp", netName = "RANDOM", outDir = ".",
  setSeed = 42L)
}
\arguments{
\item{id}{(char) vector of patient IDs}

\item{numNets}{(integer) number of networks to generate}

\item{minNodes}{(integer) min number of patients in networks}

\item{maxNodes}{(integer) max number of patients in networks}

\item{sizeProb}{(char) Probability distribution for number of patients
in a networks. Options are "unif" for uniform distribution (networks
of all sizes between minNodes and maxNodes are equally likely) sand "exp"
for exponentially decaying function (more smaller networks than larger
networks)}

\item{netName}{(char) prefix for network name.}

\item{outDir}{(char) output directory}
}
\value{
char vector of network names
}
\description{
Generates random binary interaction networks
}
\details{
Used as a control to assess the robustness of GeneMANIA to 
random cliques.
}
\examples{
data(TCGA_mini); 
x <- makeRandomNetworks(pheno$ID, numNets=10L,
outDir="/tmp")
}
