% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapNamedRangesToSets.R
\name{mapNamedRangesToSets}
\alias{mapNamedRangesToSets}
\title{Map named ranges to corresponding set of named ranges}
\usage{
mapNamedRangesToSets(gr, rangeList, verbose = FALSE)
}
\arguments{
\item{gr}{(GRanges) named ranged to be grouped}

\item{rangeList}{(list) sets of range names}

\item{verbose}{(logical) print detailed messages}
}
\value{
RangeList. keys are names of \code{rangeList}, values are GRanges
}
\description{
Map named ranges to corresponding set of named ranges
}
\details{
Example application is when we have named ranges each
corresponding to genes or regulatory elements, and we wish to group
these ranges based on metabolic pathway.
}

