% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_getIJ.R
\name{matrix_getIJ}
\alias{matrix_getIJ}
\title{Converts matrix index (1 to m*n) to row (m) and column (n) number}
\usage{
matrix_getIJ(dimMat, idx)
}
\arguments{
\item{dimMat}{(integer vector of length 2) output of \code{dim()} for
matrix in question}

\item{idx}{(integer vector of length n) matrix indices}
}
\value{
(matrix) n-by-2, first column has row indices ; second column
has col indices
}
\description{
Converts matrix index (1 to m*n) to row (m) and column (n) number
}

