\name{MB.pheno}
\alias{MB.pheno}
\docType{data}
\title{
Sample metadata table for medulloblastoma dataset.

}
\description{
data.frame with patient ID and tumour subtype (STATUS)
}
\usage{data(MB.pheno)}
\source{
Northcott et al. (2011). J Clin Oncol. 29 (11):1408.
}
\references{
Northcott et al. (2011). J Clin Oncol. 29 (11):1408.
}
\examples{
data(MB.pheno)
head(MB.pheno)
}
\keyword{datasets}
