\name{modelres}
\alias{modelres}
\docType{data}
\title{
	Sample output of getResults()
}
\description{
Output of getResults() generated by running toymodel.
toymodel is itself the output of buildPredictor() run on a simple dataset for binary breast tumour classification using two genomic data sources.
BRCA data were downloaded using curatedTCGAData for mRNA and miRNA expression. buildPredictor()] was run by scoring features out of 2, with selected features passing 1 out of 2.
Tumours were labelled either "Luminal.A" or "other".
See details of getResults() for output format.
}
\usage{data(modelres)}
\examples{
data(modelres)
head(modelres)
}
\keyword{datasets}
