% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveInteractionNets.R
\name{moveInteractionNets}
\alias{moveInteractionNets}
\title{moves interaction networks when compiling database for sparse genetic
workflow}
\usage{
moveInteractionNets(netDir, outDir, pheno, fileSfx = "_cont.txt")
}
\arguments{
\item{netDir}{(char) source directory}

\item{outDir}{(char) target directory}

\item{pheno}{(data.frame) contains patient ID and STATUS}

\item{fileSfx}{(char) suffix to strip from network file names before
registering in metadata tables}
}
\value{
No value. Side effect of moving interaction nets to target
directory and creating network-related metadata files used to compile
feature database
}
\description{
moves interaction networks when compiling database for sparse genetic
workflow
}
