% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normDiff.R
\name{normDiff}
\alias{normDiff}
\title{Similarity metric of normalized difference}
\usage{
normDiff(x)
}
\arguments{
\item{(numeric)}{vector of values, one per patient (e.g. ages)}
}
\description{
Similarity metric of normalized difference
}
\details{
Similarity metric used when data for a network consists of
exactly 1 continuous variable  (e.g. a network based only on "age"). 
When number of variables is 2-5, use avgNormDiff() which 
takes the average of normalized difference for individual variables
}
