\name{pathway_GR}
\alias{pathway_GR}
\docType{data}
\title{
	List of genomic ranges mapped to pathways
}
\description{
List object. Keys are pathway names, values are GRanges objects with coordinates of corresponding genes. Small subset of pathways sufficient for package examples.
}
\usage{data(pathway_GR)}
\examples{
data(pathway_GR)
head(pathway_GR)
}
\keyword{datasets}
