\name{pathwayList}
\alias{pathwayList}
\docType{data}
\title{
	Sample list of pathways
}
\description{
List where keys are pathway names and values are character vectors comprising of member genes for corresponding pathways
}
\usage{data(pathwayList)}
\examples{
data(pathwayList)
head(pathwayList)
}
\keyword{datasets}
