% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfCalc.R
\name{perfCalc}
\alias{perfCalc}
\title{Computes variety of predictor evaluation measures based on the confusion
matrix}
\usage{
perfCalc(dat)
}
\arguments{
\item{dat}{(data.frame): 4 columns: tp, fp, tn, fn. One row per cutoff
score for feature selection}
}
\value{
(list)
stats (data.frame): score, f1, ppv, precision and recall. One row
per cutoff for feature selection
auc (numeric between 0 and 1): AUC of overall ROC curve
prauc (numeric between 0 and 1): AUC of overall precision-recall curve
}
\description{
Computes variety of predictor evaluation measures based on the confusion
matrix
}

