% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfCalc_multiClass.R
\name{perfCalc_multiClass}
\alias{perfCalc_multiClass}
\title{Computes performance measures in multiway classification}
\usage{
perfCalc_multiClass(realLabels, predLabels)
}
\arguments{
\item{realLabels}{(char) vector of real data labels}

\item{predLabels}{(char) vector of predicted data labels}
}
\value{
(matrix) (N+1) named rows, one per class and one for average. 
Columns contain: tp,tn,fp,fn,ppv,recall,f1,accuracy
}
\description{
Computes performance measures in multiway classification
}
\details{
First converts problem of N-way classification into N
binary classification problems ("A/not-A","B/not-B",etc.,).
Then computes precision/recall measure for each class. Finally returns
measures for each class as well as an overall average measure across
all binary.
}
