\name{pheno}
\alias{pheno}
\docType{data}
\title{
Sample metadata table
}
\description{
data.frame with patient ID (ID), sample type (Type), tumour subtype (STATUS).
From TCGA 2012 breast cancer paper (see reference).
}
\usage{data(pheno)}
\source{
The Cancer Genome Atlas. (2012). Nature 490:61-70.
}
\references{
The Cancer Genome Atlas. (2012). Nature 490:61-70.
}
\examples{
data(pheno)
head(pheno)
}
\keyword{datasets}
