\name{pheno_full}
\alias{pheno_full}
\docType{data}
\title{
Subsample of TCGA breast cancer data used for netDx function examples
}
\description{
Patient ID and tumour status in "pheno", subsample of gene expression in "xpr" and CNV data in "cnv_GR"
}
\usage{data(pheno_full)}
\source{
The Cancer Genome Atlas. (2012). Nature 490:61-70.
}
\references{
The Cancer Genome Atlas. (2012). Nature 490:61-70.
}
\examples{
data(pheno_full)
head(pheno_full)
}
\keyword{datasets}
