% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tSNE.R
\name{plot_tSNE}
\alias{plot_tSNE}
\title{Plot tSNE}
\usage{
plot_tSNE(psn, pheno)
}
\arguments{
\item{psn}{(matrix) Patient similarity network represented as adjacency
matrix (symmetric). Row and column names are patient IDs. Note that NA
values will be replaced by very small number (effectively zero).}

\item{pheno}{(data.frame) Patient labels. ID column is patient ID and 
STATUS is patient label of interest. tSNE will colour-code nodes by 
patient label.}
}
\value{
(Rtsne) output of Rtsne call. Side effect of tSNE plot
}
\description{
Plot tSNE
}
\details{
Plots tSNE of patient similarity network using Rtsne
}
