% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllResults.R
\name{plotAllResults}
\alias{plotAllResults}
\title{Wrapper to plot all predictor results}
\usage{
plotAllResults(pheno, inDir, outDir, fsCutoff = 10L, fsPctPass = 0.7,
  pathwaySet = NULL, eMap_colorScheme = "cont_heatmap", eMap_min = 1,
  eMap_max = 10, setName = "predictor")
}
\arguments{
\item{pheno}{(data.frame) sample metadata. Must have ID and STATUS column}

\item{inDir}{(char) main directory with results from nested CV run}

\item{outDir}{(char) directory to store plots and results}

\item{fsCutoff}{(integer) net must score >= this in CV score cutoff to be
called feature-selected}

\item{fsPctPass}{(numeric from 0 to 1) fraction of splits for which net 
must score >= fsCutoff to be called feature-selected}

\item{pathwaySet}{(list) set of grouped units eligible for EnrichmentMap view
If set to NULL, no EnrichmentMap plotted. If provided, must be limited to 
the units measured in the dataset. For example, pathway definitions should
contain only the genes measured in the input data, and not all genes.}

\item{eMap_colorScheme}{(char) color scheme for node fills in the Enrichment
map. See plotEmap::colorScheme()}

\item{eMap_min}{(int) min feature score to include in enrichment map}

\item{eMap_max}{(int) max feature score to include in enrichment map}

\item{setName}{(char) name of analysis. Used in plot titles}
}
\value{
(list) output from all individual result scripts called.
1) predPerf: output of plotPerf()
2) featScores:
3) featSelNet: (list) keys are classes, values are vectors of feature-
selected nets
4) Emap_input: (list) keys are classes, values are output of 
writeEMapInput_many(). Only if pathwaySet is provided. 
5) EMap_pngFiles: (list) keys are classes, values are path to png files
generated by plotting the EnrichmentMap. Output of plotEMap()
6) PSN: Output of plotIntegratedPSN()
}
\description{
Wrapper to plot all predictor results
}
\details{
only works for binary classification. Assumes the output directory
structure generated by running runPredictor_nestedCV()
}
