% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmap.R
\name{plotEmap}
\alias{plotEmap}
\title{Using GMT files, generate enrichment maps and apply Auto annotate app to them}
\usage{
plotEmap(gmtFile, nodeAttrFile, netName = "generic", minScore = 1,
  maxScore = 10)
}
\arguments{
\item{gmtFile}{(char) file path to GMT file (generated by writeEMap_file()).
NOTE: This needs to be the absolute path name}

\item{nodeAttrFile}{(list) file path to nodeAttr.txt file (generated by 
writeEMap_file())}
}
\description{
Using GMT files, generate enrichment maps and apply Auto annotate app to them
}
\examples{
plotEmap(gmt_file, node_attrFile, netName="HighRisk")
}
