% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIntegratedPatientNetwork.R
\name{plotIntegratedPatientNetwork}
\alias{plotIntegratedPatientNetwork}
\title{Visualize integrated patient similarity network based on selected features}
\usage{
plotIntegratedPatientNetwork(dataList, groupList, makeNetFunc,
  setName = "predictor", topX = 0.05, aggFun = "MAX",
  calcShortestPath = FALSE, showStats = FALSE, outDir = tempdir(),
  numCores = 1L, nodeSize = 50L, edgeTransparency = 40L,
  nodeTransparency = 155L, plotCytoscape = FALSE, verbose = FALSE)
}
\arguments{
\item{dataList}{(MultiAssayExperiment) patient data & labels used as input}

\item{groupList}{(list) features to use to create integrated patient network.
Identical in structure to groupList in buildPredictor() method. This is a 
list of lists, where the outer list corresponds to assay (e.g. mRNA,
clinical) and inner list to features to generate from that datatype.}

\item{makeNetFunc}{(function) function to create features}

\item{setName}{(char) name to assign the network in Cytoscape}

\item{topX}{(numeric between 0 and 1) fraction of strongest edges to keep
e.g. topX=0.2 will keep 20\% top edges}

\item{aggFun}{(char) function to aggregate edges from different PSN}

\item{calcShortestPath}{(logical) if TRUE, computes weighted shortest path
Unless you plan to analyse these separately from looking at the shortest 
path violin plots or integrated PSN in Cytoscape, probably good to set to 
FALSE.}

\item{showStats}{(logical) if FALSE, suppresses shortest path-related 
stats, such as one-sided WMW test for testing shorter intra-class distances}

\item{outDir}{(char) path to directory for intermediate files. Useful for
debugging.}

\item{numCores}{(integer) number of cores for parallel processing}

\item{nodeSize}{(integer) size of nodes in Cytoscape}

\item{edgeTransparency}{(integer) Edge transparency. 
Value between 0 and 255, with higher numbers leading to more opacity.}

\item{nodeTransparency}{(integer) Node transparency.
Value between 0 and 255, with higher numbers leading to more opacity.}

\item{plotCytoscape}{(logical) If TRUE, plots network in Cytoscape.
Requires Cytoscape software to be installed and running on the computer
when the function call is being made.}

\item{verbose}{(logical) print detailed messages}
}
\value{
(list) information about the integrated network
similarity network
2) patientDistNetwork_pruned (matrix) the network plotted in
Cytoscape. Also note that this is a dissimilarity network, 
so that more similar nodes have smaller edge weights
3) colLegend (data.frame): legend for the patient network
plotted in Cytoscape. Columns are node labels (STATUS) and
colours (colour)
4) topX (numeric) value of topX parameter
5) aggFun (char) function used to aggregate nets
6) outDir (char) value of outDir parameter
}
\description{
Visualize integrated patient similarity network based on selected features
}
\details{
Generates a Cytoscape network where nodes are patients and edges
are weighted by aggregate pairwise patient similarity.
Integrated PSN plotting is intended to run
after feature selection, which identifies the subset of input networks
predictive for each class of interest. The method of generating the network
is as follows:
All networks feature-selected in either patient ategory are concatenated;
where a network is feature-selected in both categories, it is included once.
The similarity between two patients in the integrated network is the mean of
corresponding pairwise similarities. Dissimilarity is defined as
1-similarity, and Dijkstra distances are computed on this resulting network.
For visualization, only edges representing the top fraction of distances
(strongest edge weights) are included.
}
