% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIntegratedPSN.R
\name{plotIntegratedPSN}
\alias{plotIntegratedPSN}
\title{Visualize integrated patient similarity network based on selected features}
\usage{
plotIntegratedPSN(setName = "predictor", pheno, netDir, topX = 0.05,
  aggFun = "MAX", outDir = ".", calcShortestPath = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{pheno}{(data.frame) sample metadata, requires ID and STATUS column.
The status column will be used to retrieve patient class names}

\item{netDir}{(char) path to directory with networks}

\item{topX}{(numeric between 0 and 1) fraction of strongest edges to keep
e.g. topX=0.2 will keep 20\% top edges}

\item{aggFun}{(char) function to aggregate edges from different PSN}

\item{outDir}{(char) path to directory for tmp work}

\item{calcShortestPath}{(logical) if TRUE, computes weighted shortest path
Unless you plan to analyse these separately from looking at the shortest 
path violin plots or integrated PSN in Cytoscape, probably good to set to 
FALSE.
for all pairwise classes}

\item{verbose}{(logical) print detailed messages}
}
\value{
(list) information about the integrated network
is the unpruned network and not the one displayed in Cytoscape. Edges are
weighted by similarity
2) aggPDN_pruned: (char) path to file with pruned patient dissimilarity
network, derived by inverting edge weights of the net in aggPSN_Full and
keeping the strongest edges.
2) incNets: (char) vector of networks used for integration
3) network_suid: (char) Cytoscape network identifier, so user may further
4) netPng: (char) path to png file with patient dissimilarity network
created by Cytoscape
}
\description{
Visualize integrated patient similarity network based on selected features
}
\details{
Generates a Cytoscape network where nodes are patients and edges
are weighted by aggregate pairwise patient similarity.
Integrated PSN plotting is intended to run
after feature selection, which identifies the subset of input networks
predictive for each class of interest. The method of generating the network
is as follows:
All networks feature-selected in either patient ategory are concatenated;
where a network is feature-selected in both categories, it is included once.
The similarity between two patients in the integrated network is the mean of
corresponding pairwise similarities. Dissimilarity is defined as
1-similarity, and Dijkstra distances are computed on this resulting network.
For visualization, only edges representing the top fraction of distances
(strongest edge weights) are included.
}
