% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_diagnostics.R
\name{plotNetEdgeDistribution}
\alias{plotNetEdgeDistribution}
\title{plot distribution of PSN edge weights}
\usage{
plotNetEdgeDistribution(netDir, setName = "netDir",
  randomSample = NULL, grepFiles = "cont.txt",
  stripFromFileName = "_cont.txt", numPatients = NULL,
  verbose = TRUE)
}
\arguments{
\item{netDir}{(char) path to directory of networks. Networks must be
in SIF format; i.e. with three columns: source, target, edge weight.}

\item{setName}{(char) used for output file name (<setName>_edgeDistr.pdf)}

\item{randomSample}{(integer) if set to NULL. Plots for all networks.
If set to an integer plots the number of networks requested, to the max
number of networks available}

\item{grepFiles}{(char) searches for files with the provided string.}

\item{stripFromFileName}{(char) vector of char patterns to remove from file name}

\item{numPatients}{(integer) number of patients expected in network.
Used to calculate sparsity. If set to NULL, sparsity will not be calculated.}

\item{verbose}{(logical) print messages}
}
\value{
No value. Side effect of creating plots of network edge-weight 
distribution and printing network statistics to console. Files created
include:
1) <netDir>/<setName>_edgeDistr.pdf: Violin plots showing edge-weight
distribution for each network
2) <netDir>/<setName>_edgeCount.pdf: Single violin plot of edge count
distribution across all networks.
The script also prints the mean and SD of edge weight across all networks
}
\description{
plot distribution of PSN edge weights
}
\details{
Used for comparing the properties of PSN for different similarity
metrics
}
