% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPerf.R
\name{plotPerf}
\alias{plotPerf}
\title{Plots various measures of predictor performance for binary classifiers}
\usage{
plotPerf(resList = NULL, inFiles, predClasses)
}
\arguments{
\item{resList}{(list) list of prediction results. If provided, the method
will ignore inDir}

\item{inFiles}{(char) path to predictionResults.txt files.
A vector, each with absolute paths to predictionResults.txt}

\item{predClasses}{(char) vector of class names.}
}
\value{
(list) each key corresponds to an input file in inDir.
Value is a list with:
1) stats: 'stats' component of perfCalc
2) rocCurve: ROCR performance object for ROC curve
3) prCurve: ROCR performance object for PR curve
4) auroc: Area under ROC curve
5) aupr: Area under PR curve
6) accuracy: Accuracy

Side effect of plotting in a 2x2 format:
1) mean+/-SEM AUROC
2) mean+/-SEM AUPR
3) ROC curve for all runs plus average
4) PR curve for all runs plus average
}
\description{
Plots various measures of predictor performance for binary classifiers
}
\details{
Plots individual and average ROC/PR curves. mean+/-SEM performance 
for a predictor run using nested
cross-validation or a similar repeated design.
predictionResults.txt contains a (data.frame)
}
\examples{
inDir <- sprintf('\%s/extdata/example_output',
		path.package('netDx'))
inFiles <- paste(inDir, sprintf("rng\%i",1:3),"predictionResults.txt",
 sep="/")
plotPerf(inFiles, predClasses = c('LumA','notLumA'))
}
