% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPerf_multi.R
\name{plotPerf_multi}
\alias{plotPerf_multi}
\title{Plots a set of ROC/PR curves with average.}
\usage{
plotPerf_multi(inList, plotTitle = "performance", plotType = "ROC",
  xlab = "TPR", ylab = "FPR", meanCol = "darkblue", xlim = c(0, 1),
  ylim = c(0, 1))
}
\arguments{
\item{inList}{(list) ROCR::performance objects, one per iteration}

\item{plotType}{(char) one of ROC | PR | custom. Affects x/y labels}

\item{xlab}{(char) x-axis label}

\item{ylab}{(char) y-axis label}

\item{meanCol}{(char) colour for mean trendline}
}
\value{
No value. Side effect of plotting ROC and PR curves
}
\description{
Plots a set of ROC/PR curves with average.
}
\details{
Plots average curves with individual curves imposed.
}
