% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{predict patient labels}
\usage{
predict(
  trainMAE,
  testMAE,
  groupList,
  featSel,
  makeNetFunc,
  outDir,
  impute = FALSE,
  verbose = FALSE,
  numCores = 1L,
  JavaMemory = 4L,
  debugMode = FALSE
)
}
\arguments{
\item{trainMAE}{(MultiAssayExperiment) patient data for training samples. Same as provided to buildPredictor()}

\item{testMAE}{(MultiAssayExperiment) new patient dataset for testing model. Assays must be the same as for trainMAE.}

\item{groupList}{(list) list of features used to train the model. Keys are data types, and values are lists for groupings within those datatypes.
e.g. keys could include {'clinical','rna','methylation'}, and values within 'rna' could include pathway names {'cell cycle', 'DNA repair'}, etc.,
featSel will be used to subset}

\item{featSel}{(list) selected features to be used in the predictive model. 
keys are patient labels (e.g. "responder/nonresponder"), and values are feature names 
identified by running buildPredictor(). Feature names must correspond to names of groupList, from which they will be subset.}

\item{makeNetFunc}{(function) function to create PSN features from patient data. See makeNetFunc in buildPredictor() for details}

\item{outDir}{(char) directory for results}

\item{impute}{(logical) if TRUE imputes train and test samples separately before creating features. Currently unsupported.}

\item{verbose}{(logical) print messages}

\item{numCores}{(integer) number of CPU cores for parallel processing}

\item{JavaMemory}{(integer) memory in (Gb) used for each fold of CV}

\item{debugMode}{(logical) Set to TRUE for detailed messages. Used for debugging.}
}
\value{
(data.frame) predicted patient similarities and labels
columns are:  1) ID, 2) STATUS (ground truth), 3) <label>_SCORE: similarity score for the corresponding label,
4) PRED_CLASS: predicted class
}
\description{
Once a model is trained, this function is used to classify new patients using selected features
}
