% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooMutationPropagation.R
\name{prop_m}
\alias{prop_m}
\title{This function applies the random walk with restart propagation algorithm to a
matrix of patients profiles}
\usage{
prop_m(mat, net, cl, no_cores)
}
\arguments{
\item{mat}{(data.frame) sparse matrix of patient profiles. Rownames
are unique genes. Colnames are unique patients. A cell is a numeric value.}

\item{net}{(data.frame) adjancy matrix format of a network}

\item{cl}{(SOCKcluster) cluster object created with makeCluster function 
from parallel}

\item{no_cores}{(numeric) number of cores used to create the cluster object}
}
\value{
(data.frame) continuous matrix of patient profiles in which each gene
  has the final propagation score
}
\description{
This function applies the random walk with restart propagation algorithm to a
matrix of patients profiles
}
\details{
A network is an undirected graph G defined by a set of nodes
  corresponding to genes, and edges connecting nodes with an experimental
  evidence of interaction. A priori nodes are genes for which an information
  is known. A novel node is a candidate for being associated to the nodes
  above based on their information. A node prediction task leads to detect
  novel nodes and propagation techniques are largely applied for the purpose.
  Network-based propagation algorithms for node prediction transfer the
  information from a priori nodes to any other node in a network. Each node
  gets an imputation value which assesses how much information got. The
  prediction is based on the guilty-by-association principle. A node with a
  high imputation value has a high probability to be associated to a priori
  nodes. E.g. in a house where room A has one heater, if room B is the second
  hottest room it means that B is close to A and that there is a high
  probability that they share a door or wall. These algorithms exploit the
  global topology of the network. However, when they are applied to detect if
  unknown nodes are functionally associated to known ones, they may suffer of
  a drawback depending by the context. In biology, two functionally related
  fragments interact physically (direct interaction) or interact indirectly
  thanks to one or very few mediators. Therefore, exploring too far
  similarities between nodes can introduce noise in the prediction. We apply
  a random walk with restart propagation algorithm which resolution is set to
  0.2 for giving high values only to the close neighbours of the a priori
  nodes.
}
