% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneNetByStrongest.R
\name{pruneNet}
\alias{pruneNet}
\title{Prune network by retaining strongest edges}
\usage{
pruneNet(net, vertices, pctX = 0.1, useTop = TRUE)
}
\arguments{
\item{net}{(data.frame)  Network to prune. Columns are: source,target,weight}

\item{vertices}{(char) node names. Should match those in net[,1:2]}

\item{pctX}{(numeric 0 to 1) Fraction of top/bottom edges to retain}

\item{useTop}{(logical) if TRUE prunes to top pctX edges; else
prunes to bottom pctX edges}
}
\value{
(data.frame) pruned network. Three columns: AliasA, AliasB, and 
weight
}
\description{
Prune network by retaining strongest edges
}
