% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneNetByStrongest.R
\name{pruneNetByStrongest}
\alias{pruneNetByStrongest}
\title{Prune network by retaining strongest edges}
\usage{
pruneNetByStrongest(net, vertices, topX = 0.1)
}
\arguments{
\item{net}{(data.frame)  Network to prune. Columns are: source,target,weight}

\item{vertices}{(char) node names. Should match those in net[,1:2]}

\item{topX}{(numeric 0 to 1) Fraction of strongest edges to retain}
}
\value{
(data.frame) pruned network. Three columns: AliasA, AliasB, and 
weight
}
\description{
Prune network by retaining strongest edges
}

