% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneNets.R
\name{pruneNets}
\alias{pruneNets}
\title{Prune interaction networks to keep only the networks and patients 
requested}
\usage{
pruneNets(oldDir, newDir = tempdir(), filterNets = "*",
  filterIDs = "*", netSfx = "_cont.txt$", verbose = TRUE)
}
\arguments{
\item{oldDir}{(char) path to directory with original networks}

\item{newDir}{(char) path to output directory for pruned networks}

\item{filterNets}{(char) vector of networks to include. These should 
match filenames in \code{netDir}. Value of '*' results in pruning all
networks}

\item{filterIDs}{(char) patients to include in pruned networks. These
should match nodes in the input interaction networks}

\item{netSfx}{(char) suffix for network file names. Only used if 
\code{filterNets='*'}.}

\item{verbose}{(logical) print messages}
}
\value{
(no value). Side effect of writing pruned network files to 
\code{newDir}
}
\description{
Prune interaction networks to keep only the networks and patients 
requested
}
\details{
This function is crucial for patient data that is highly 
sparse; examples include patient CNVs indels, as opposed to full matrix
measures (gene expression, questionnaire data). Each step where the pool
of patients is subset - e.g. limiting feature selection only to patients
in training set - changes the set of networks that are eligible. 
Some networks may only contain test patients, while others may contain
a single edge between a training and a test patient. Upon subsetting,
such networks are no longer eligible for downstream use, such as 
feature selection. This function rewrites those subnetworks of the 
original networks that consist of eligible patients.
}
\examples{
data(npheno)
netDir <- sprintf('\%s/extdata/example_nets',path.package('netDx'))
pruneNets(netDir,'~/tmp',filterIDs=npheno[seq_len(10),],
 netSfx='txt$')
}
