% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{psn__builtIn}
\alias{psn__builtIn}
\title{make PSN for built-in similarity functions}
\usage{
psn__builtIn(settings, ...)
}
\arguments{
\item{settings}{(list) from makeNetFunc}
}
\description{
make PSN for built-in similarity functions
}
