% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{psn__corr}
\alias{psn__corr}
\title{wrapper for PSNs using Pearson correlation}
\usage{
psn__corr(settings, verbose, ...)
}
\arguments{
\item{settings}{(list) from makeNetFunc}

\item{verbose}{(logical) print messages}

\item{...}{parameters for makePSN_NamedMatrix()}
}
\value{
(char) names of networks created. Side effect of network creation.
}
\description{
wrapper for PSNs using Pearson correlation
}
