% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{randAlphanumString}
\alias{randAlphanumString}
\title{Generate random alphanumerical string of length 10}
\usage{
randAlphanumString(numStrings = 1L)
}
\arguments{
\item{numStrings}{(integer) number of strings to generate}
}
\value{
vector of length n, each with 10-char alphanumerical strings
}
\description{
Generate random alphanumerical string of length 10
}
\details{
Used to create multiple temporary directories during an R session
}
\examples{
randAlphanumString()
}
