% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPathwayFile.R
\name{readPathways}
\alias{readPathways}
\title{Parse GMT file and return pathways as list}
\usage{
readPathways(fname, MIN_SIZE = 10L, MAX_SIZE = 500L, EXCLUDE_KEGG = TRUE,
  IDasName = FALSE, verbose = TRUE, getOrigNames = FALSE)
}
\arguments{
\item{fname}{(char) path to pathway file in gmt format
pathway score to include pathway in the filter list}

\item{MIN_SIZE}{(integer) min num genes allowed in a pathway. Pathways
with fewer number of genes are excluded from the output list}

\item{MAX_SIZE}{(integer) max num genes allowed in a pathway. Pathways
with gene counts greater than this are excluded from the output list}

\item{getOrigNames}{(logical) when TRUE also returns a mapping of the
cleaned pathway names to the original names}

\item{EXCLUDE_KEGG:}{(boolean) If TRUE exclude KEGG pathways. Our
experience has been that some KEGG gene sets are to broad to be 
physiologically relevant}

\item{IDasName:}{(boolean) Value for key in output list. 
If TRUE, uses db name and ID as name (e.g.  KEGG:hsa04940)
If FALSE, pathway name. If TRUE,}
}
\value{
Depends on value of getOrigNames. If FALSE (Default), list with
pathway name as key, vector of genes as value. If TRUE, returns list of
length two, (1) geneSets: pathway-gene mappings as default, 
(2) pNames: data.frame with original and cleaned names.
}
\description{
Parse GMT file and return pathways as list
}
\details{
The GMT file format currently supported should match the ones
found at http://downloads.baderlab.org. The original GMT file format is:
<set name><set description><member 1><member 2>...<member N>, 
one row per set, with values tab-delimited.
The version at baderlab.org has additional unique formatting of the
<set name> column as follows:
<pathway_full_name>%<pathway_source>%<pathway_source_id>

Example:
UREA CYCLE%HUMANCYC%PWY-4984	urea cycle	ASS1	ARG1	CPS1	ASL	OTC
ABACAVIR METABOLISM%REACTOME DATABASE ID RELEASE 55%2161541	Abacavir metabolism	ADH1A	GUK1	ADAL	PCK1	NT5C2

This function requires the specific formatting of the first column
to assign the key name of the output list (see \code{useIDasName} 
argument).
}
\examples{
pathFile <- sprintf("\%s/extdata/pathways.gmt",
	path.package("netDx"))
pathwayList    <- readPathways(pathFile)

}

