% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileFeatures.R
\name{replacePattern}
\alias{replacePattern}
\title{Replace pattern in all files in dir}
\usage{
replacePattern(pattern = ",", target = ".", path = getwd(), fileType = "txt$")
}
\arguments{
\item{pattern}{(char) pattern to find}

\item{target}{(char) pattern to replace}

\item{path}{(char) dir to replace pattern in}

\item{fileType}{(char) pattern for files to replace pattern in}
}
\value{
No value. Files have patterns replaced in place.
}
\description{
find/replace pattern in all files of specified file type
in specified directory. Needed to modify number format when intefacing
with GeneMANIA, on  French locale machines. Without this step,
CacheBuilder throws error with commas.
}
