% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_modelEval.R
\name{resampling_modelEval}
\alias{resampling_modelEval}
\title{Resampling-based predictor, evaluate test samples}
\usage{
resampling_modelEval(pheno, pdat, predClass, unitSets, p_GR = NULL,
  unitSet_GR = NULL, bestCutoff, dataDir, outDir, netScores = NULL,
  numCores)
}
\arguments{
\item{pheno}{(data.frame) patient metadata. ID, class labels (STATUS),
and whether train/test status (TT_STATUS): a patient is query if "TRAIN",
else is "TEST"}

\item{pdat}{(data.frame) patient data. Rows are unit variables, columns 
are patients. Rows and columns must be named.}

\item{predClass}{(char) class of interest to predict.}

\item{unitSets}{(list) groupings of unit variables; each grouping will
be converted into its own PSN.}

\item{p_GR}{(GRanges) GRanges of patient CNVs. Has ID column in
metadata, containing patient IDs}

\item{unitSet_GR}{(list) sets of GRanges to group CNVs (e.g.
could have one GRanges per pathway, corresponding to regions in that 
pathway}

\item{bestCutoff}{(integer) nets with score >= bestCutoff are used in 
the final model}

\item{dataDir}{(char) dir containing subdirs corresponding to each 
subtype}

\item{outDir}{(char) directory in which results must be written}

\item{netScores}{(list) one key per subtype, and value is a data.frame
with net name (column 1) and score (column 2). If NULL, is obtained
using dataDir}

\item{numCores}{(integer) number of cores for parallel processing}

\item{patNets}{(list of chars) for each class (key), subset of "unitSets"
for which patient nets must be created. e.g. feature-selected networks
for the corresponding patient subtype.}
}
\value{
(list)
1. test_res: output of GM_predClass_once
2. netScores (list) keys are subtypes, values are feature selected
nets
}
\description{
Resampling-based predictor, evaluate test samples
}
