% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_pickBestCutoff.R
\name{resampling_pickBestCutoff}
\alias{resampling_pickBestCutoff}
\title{Pick best cutoff following resampling of training data}
\usage{
resampling_pickBestCutoff(perf_resamp)
}
\arguments{
\item{perf_resamp}{(list) the output of GM_predClass_cutoffs(), 
with one entry for each round of resampling}
}
\value{
(list) with performance at each cutoff:
1) perfmat (matrix): tpr, fpr, accuracy,ppv for each resampling. One 
row per cutoff, 4 columns per round of resampling.
2) bestCutoff (integer): cutoff with the highest accuracy
3) mu (matrix): average of tpr, fpr, accuracy, ppv across resamplings.
one row per cutoff.
}
\description{
Pick best cutoff following resampling of training data
}
