% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_pickBestCutoff_CNV.R
\name{resampling_pickBestCutoff_CNV}
\alias{resampling_pickBestCutoff_CNV}
\title{pick best cutoff for CNV-based predictor}
\usage{
resampling_pickBestCutoff_CNV(resampPerf)
}
\arguments{
\item{resampPerf}{(matrix) output of RR_featureTally stored in 
<outDir>/resampPerf.Rdata. Has one row per score, and (7*resamp) columns. 
7 columns for each resampling round. see RR_featureTally() for more details.
scores must be in rownames(resampPerf)}
}
\value{
(list)
1) bestCutoff (integer) score with highest accuracy
3) accuracy (matrix) accuracy for each score cutoff
}
\description{
pick best cutoff for CNV-based predictor
}
