% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_predTest_CNV.R
\name{resampling_predTest_CNV}
\alias{resampling_predTest_CNV}
\title{evaluate test performance for CNV-based predictor}
\usage{
resampling_predTest_CNV(testPheno, test_p_GR, selFeature_GR, outDir)
}
\arguments{
\item{testPheno}{(data.frame) sample metadata table with ID and STATUS}

\item{test_p_GR}{(GRanges) genomic events in patients. Must have ID column
that maps an event to a patient, and LOCUS_NAMES with comma-separated names
of units (e.g. genes) that the ranges overlap}

\item{selFeature_GR}{(list of GRanges) genomic ranges for selected features
NOTE: These are the features over which patients will be evaluated so 
limit to nets that pass feature selection}

\item{outDir}{(char) root directory in which data must be saved}
}
\description{
evaluate test performance for CNV-based predictor
}

