% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR_featureTally.R
\name{RR_featureTally}
\alias{RR_featureTally}
\title{Computes positive and negative calls upon changing stringency of 
feature selected networks (binary networks only)}
\usage{
RR_featureTally(netmat, phenoDF, TT_STATUS, predClass, pScore, outDir,
  cliqueFilter = TRUE, cliqueNets, maxScore = 30L)
}
\arguments{
\item{netmat}{(matrix) output of countPatientsInNet. Should contain
all patients in dataset that overlap 1+ network}

\item{phenoDF}{(data.frame) patient ID and STATUS}

\item{TT_STATUS}{(list) output of splitTestTrain_partition; should be
same as used for cross validation}

\item{predClass}{(char) class to be predicted}

\item{pScore}{(list of data.frames) contains 10-fold CV score, one entry
for each resampling of the data. The data.frame has two columns: 
1) pathway name, 2) pathway score}

\item{outDir}{(char) path to dir where results should be written}

\item{cliqueFilter}{(logical) was clique filtering used?}

\item{cliqueNets}{(list of chars) networks passing clique filtering}

\item{maxScore}{(integer) max achievable score for pathways
corresponding to N-way resampling}
}
\value{
No value. Side effect of writing the following files: 
1) pathway tally file: \code{<outDir>/pathway_cumTally.txt} : 
pathway name, cumulative score over N-way data resampling.
2) positive,negative calls at each cutoff: 
\code{<outDir>/RR_changeNetSum_stats_denAllNets.txt}: 
network score cutoff (score); num networks at cutoff (numPathways) ; 
total +, ground truth (pred_tot); + calls (pred_ol); 
+ calls as pct of total (pred_pct); total -, ground truth (other_tot) ; 
- calls  (other_ol) ; - calls as pct of total (other_pct) ; ratio of 
pred_pct and other_pct (rr) ; min. pred_pct in all resamplings 
(pred_pct_min) ; max pred_pct in all resamplings (pred_pct_max) ; 
min other_pct in all resamplings (other_pct_min); max other_pct in all
resamplings (other_pct_max)
3) positive, negative calls at each cutoff ; clique filtering option:  
\code{<outDir>/RR_changeNetSum_stats_denCliqueNets.txt}: format same
as the previous file. However, the denominator here is limited to
patients present in networks that pass clique filtering.
}
\description{
Computes positive and negative calls upon changing stringency of 
feature selected networks (binary networks only)
}
\details{
This function computes predictor performance in the context
of binary networks, where + and - calls are based on membership (or 
lack thereof) in feature selected networks. An example would be networks
based on CNV occurrence in cellular pathways; in this use case, a +
is based on patient membership in feature-selected networks. This
function takes the output data from a feature selection exercise
and computes the number and fraction of positive and negative calls at
each level of feature selection stringency. The output of this function
can then be used to compute performance measures such as the ROC or 
precision-recall curve.
}

