% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGeneMANIA.R
\name{runGeneMANIA}
\alias{runGeneMANIA}
\title{Run a GeneMANIA query}
\usage{
runGeneMANIA(GM_db, queryFile, resDir, parseReport = TRUE, verbose = TRUE,
  GMmemory = 6L, MAX_ATTEMPTS = 3L)
}
\arguments{
\item{GM_db}{(char) path to directory with GeneMANIA generic database}

\item{queryFile}{(char) path to query file}

\item{resDir}{(char) path to output directory}

\item{parseReport}{(logical) if TRUE, parses out PRANK and NRANK portions}

\item{verbose}{(logical) print messages}

\item{GMmemory}{(integer) Memory for GeneMANIA (in Gb)}

\item{MAX_ATTEMPTS}{(integer) max num attempts to run GeneMANIA before
giving up.}
}
\value{
path to GeneMANIA query result file
of results file
}
\description{
Run a GeneMANIA query
}
\examples{
GM_db <- sprintf("\%s/extdata/GM_db", path.package("netDx"))
GM_query <- sprintf("\%s/extdata/GM_query.txt",
	path.package("netDx"))
runGeneMANIA(GM_db, GM_query,tmpDir)
}

