% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGeneMANIA3.R
\name{runGeneMANIA3}
\alias{runGeneMANIA3}
\title{Run a GeneMANIA query}
\usage{
runGeneMANIA3(GM_db, queryFiles, resDir, parseReport = FALSE,
  verbose = TRUE, GMmemory = 6L, MAX_ATTEMPTS = 3L, numCores = 1)
}
\arguments{
\item{GM_db}{(char) path to directory with GeneMANIA generic database}

\item{queryFiles}{(list(char)) paths to query files}

\item{resDir}{(char) path to output directory}

\item{parseReport}{(logical) if TRUE, parses out PRANK and NRANK portions}

\item{verbose}{(logical) print messages}

\item{GMmemory}{(integer) Memory for GeneMANIA (in Gb) - a total of 
numCores*GMmemory will be used and distributed for all GM threads}

\item{numCores}{(integer) number of CPU cores for parallel processing}
}
\value{
path to GeneMANIA query result file
of results file
}
\description{
Run a GeneMANIA query
}
\examples{
GM_db <- sprintf("\%s/extdata/GM_db", path.package("netDx"))
GM_query <- sprintf("\%s/extdata/GM_query.txt",
	path.package("netDx"))
runGeneMANIA(GM_db, GM_query,"/tmp")
}
