% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{runOverallFeatureSelection}
\alias{runOverallFeatureSelection}
\title{wrapper to call selected features}
\usage{
runOverallFeatureSelection(
  featScores,
  featureSelCutoff,
  featureSelPct,
  printRes = TRUE,
  cleanNames = TRUE
)
}
\arguments{
\item{featureSelCutoff}{(integer) cutoff score for feature selection.
A feature must have minimum of this score for specified fraction of splits 
(see featureSelPct) to pass.}

\item{featureSelPct}{(numeric between 0 and 1) cutoff percent for feature selection.
A feature must have minimum score of featureSelCutoff for featureSelPct of 
train/test splits to pass.}

\item{cleanNames}{(logical) remove internal suffixes for human readability}
}
\value{
(list) Feature scores for all splits, plus those passing selection for overall predictor
featScores: (matrix) feature scores for each split
selectedFeatures: (list) features passing selection for each class; one key per class
}
\description{
wrapper to call selected features
}
