% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQuery.R
\name{runQuery}
\alias{runQuery}
\title{Run a GeneMANIA query}
\usage{
runQuery(dbPath, queryFile, resDir, parseReport = TRUE, verbose = TRUE,
  JavaMemory = 6L, MAX_ATTEMPTS = 3L)
}
\arguments{
\item{dbPath}{(char) path to directory with GeneMANIA generic database}

\item{queryFile}{(char) path to query file}

\item{resDir}{(char) path to output directory}

\item{parseReport}{(logical) if TRUE, parses out PRANK and NRANK portions}

\item{verbose}{(logical) print messages}

\item{JavaMemory}{(integer) Memory for GeneMANIA (in Gb)}

\item{MAX_ATTEMPTS}{(integer) max num attempts to run GeneMANIA before
giving up.}
}
\value{
path to GeneMANIA query result file
of results file
}
\description{
Run a GeneMANIA query
}
\examples{
dbPath <- sprintf("\%s/extdata/dbPath", path.package("netDx"))
GM_query <- sprintf("\%s/extdata/GM_query.txt",
	path.package("netDx"))
runQuery(dbPath, GM_query,"/tmp")
}
