% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupFeatureDB.R
\name{setupFeatureDB}
\alias{setupFeatureDB}
\title{setup database of features for feature selection}
\usage{
setupFeatureDB(pheno, prepDir)
}
\arguments{
\item{pheno}{(data.frame) patient metadata. Must contain ID column}

\item{prepDir}{(char) directory in which to setup database}
}
\value{
(data.frame) internal numerical id for patients (INTERNAL_ID) and
user-provided ID (ID)
}
\description{
Creates all the input files for the collection of features used in 
feature selection.
}
