% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{sim.pearscale}
\alias{sim.pearscale}
\title{various similarity functions
Similarity function: Pearson correlation followed by exponential scaling}
\usage{
sim.pearscale(dat, K = 20, alpha = 0.5)
}
\arguments{
\item{dat}{(data.frame) Patient data; rows are measures, columns are patients.}

\item{K}{(integer) Number of nearest neighbours to consider (K of KNN)}

\item{alpha}{(numeric) Scaling factor for exponential similarity kernel. 
Recommended range between 0.3 and 0.8.}
}
\value{
symmetric matrix of size ncol(dat) (number of patients) containing
pairwise patient similarities
}
\description{
Computes Pearson correlation between patients. A scaled 
exponential similarity kernel is used to determine edge weight. The 
exponential scaling considers the K nearest neighbours, so that 
similarities between non-neighbours is set to zero. Alpha is a hyperparameter
that determines decay rate of the exponential. For details
see Wang et al. (2014). Nature Methods 11:333.
}
\examples{
data(xpr)
sim <- sim.pearscale(xpr)
}
