% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCap.R
\name{simpleCap}
\alias{simpleCap}
\title{simple capitalization}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{(char) name}
}
\value{
(char) Changes case so start of each word is in upper-case, and
the rest is in lowercase
}
\description{
simple capitalization
}
\details{
used to format feature names so they are not in all-caps
}
