% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsify2.R
\name{sparsify2}
\alias{sparsify2}
\title{cleaner sparsification routine}
\usage{
sparsify2(W, outFile = "tmp.txt", cutoff = 0.3, maxInt = 50,
  EDGE_MAX = 1000)
}
\arguments{
\item{W}{(matrix) similarity matrix}

\item{outFile}{(char) path to file to write sparsified network}

\item{cutoff}{(numeric) edges with weight smaller than this are set to NA}

\item{maxInt}{(numeric) max num edges per node.}

\item{EDGE_MAX}{(numeric) max num edges in network}
}
\value{
writes SIF content to text file (node1,node2,edge weight)
}
\description{
cleaner sparsification routine
}
\details{
Sparsifies similarity matrix to keep strongest edges.
Sets diagonal and edges < cutoff to NA. Keeps strongest maxInt edges
per node. Ties are ignored. Keeps a max of EDGE_MAX edges in the network.
}
