% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsify3.R
\name{sparsify3}
\alias{sparsify3}
\title{cleaner sparsification routine - faster, matrix-based version}
\usage{
sparsify3(W, outFile = "tmp.txt", cutoff = 0.3, maxInt = 50,
  EDGE_MAX = Inf, includeAllNodes = TRUE, verbose = TRUE)
}
\arguments{
\item{W}{(matrix) similarity matrix}

\item{outFile}{(char) path to file to write sparsified network}

\item{cutoff}{(numeric) edges with weight smaller than this are set to NA}

\item{maxInt}{(numeric) max num edges per node.}

\item{EDGE_MAX}{(numeric) max num edges in network}

\item{includeAllNodes}{(logical) if TRUE, ensures at least one edge is present for each patient. This feature is required when sparsification excludes test patients that are required to be classified. If the sparsification rules exclude all edges for a patient and this flag is set, then the strongest edge for each missing patient is added to the net. Note that this condition results in the total number of edges potentially exceeding EDGE_MAX}

\item{verbose}{(logical) print detailed messages, useful for debugging}
}
\value{
writes SIF content to text file (node1,node2,edge weight)
}
\description{
cleaner sparsification routine - faster, matrix-based version
}
\details{
Sparsifies similarity matrix to keep strongest edges.
Sets diagonal and edges < cutoff to NA. Keeps strongest maxInt edges
per node. Ties are ignored. Keeps a max of EDGE_MAX edges in the network.
}
\examples{
m <- matrix(runif(500*500),nrow=500)
y <- sparsify2(m)
m <- matrix(runif(500*500),nrow=500)
y <- sparsify2(m)
}
