% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTestTrain_partition.R
\name{subsampleValidationData}
\alias{subsampleValidationData}
\title{Subsample a hold-out set from a larger patient dataset}
\usage{
subsampleValidationData(dataMAE, pctValidation = 0.2, verbose = TRUE)
}
\arguments{
\item{dataMAE}{(MultiAssayExperiment) patient data to be subsampled. Must have columns ID (patient ID)
and STATUS}

\item{pctValidation}{(numeric) Fraction of dataset to include in the validation set. Value from 0.05 to 0.95.}

\item{verbose}{(logical)}
}
\value{
(list) Keys are trainMAE and validationMAE. These contain corresponding MultiAssayExperiments for
training and test data
}
\description{
Subsample a hold-out set from a larger patient dataset
}
\details{
Creates a partition of data to be used for model validation after initial model building.
In netDx, buildPredictor() is used for model training, and selected features from this exercise
are used to validate a held-out dataset with the predict() function. Note that this function
identifies a random subsample, which may result in a validation sample that is not representative
of your training bias. If this method is used, please use data exploration techniques (e.g. UMAP)
to ensure that validation accuracy is not confounded by stratification.
}
