% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooMutationPropagation.R
\name{thresholdSmoothedMutations}
\alias{thresholdSmoothedMutations}
\title{Apply discretization to the matrix resulted from the propagation on the
sparse patient matrix}
\usage{
thresholdSmoothedMutations(m_prop, m_bin, name_dataset)
}
\arguments{
\item{m_bin}{(data.frame) binary somatic mutation sparse matrix. Rownames
  are unique genes. Colnames are unique patients. A cell contains a zero or a 
one.}

\item{m_profiles}{(data.frame) continous matrix of patient profiles resulted 
from applying the network-based propagation algorithm on a binary somatic 
mutation sparse matrix.}
}
\value{
(data.frame) binary somatic mutation matrix which sparsity has been 
decreased
}
\description{
Apply discretization to the matrix resulted from the propagation on the
sparse patient matrix
}
\details{
This function is included in the netDx use case which involves
  propagating the sparse matrix of patient's profiles to reduce its sparsity.
  This function applies discretization on the propagated matrix of patient
  profiles. It sets to 1 the genes which got the highest propagation value.
  While, the remaining genes are set to 0. This discretization is driven by
  the fact that higher is the propagation value and higher is the chance that
  the gene is involved in the patient condition and expression/mutation
  profile. On the contrary, genes which got either a medium or a low value
  are not trustable.
}
\examples{
  set.seed(8)
  numCores <- 8L
  library("netDx")
  require("MultiAssayExperiment")
  outDir <- "/output"
  out_plot_perf <- paste(outDir,"plot_performances.png",sep=getFileSep())
  out_features_table <- paste(outDir,"OV_features_table.csv",sep=getFileSep())
  out_res_rda <- paste(outDir,"OV_results.rda",sep=getFileSep())
  genoFile <- paste(path.package("netDx"),"extdata","OV_mutSmooth_geno.txt",
  sep=getFileSep())
  geno <- read.delim(genoFile,sep="\t",header=TRUE,as.is=TRUE)
  phenoFile <- paste(path.package("netDx"),"extdata","OV_mutSmooth_pheno.txt",
  sep=getFileSep())
  pheno <- read.delim(phenoFile,sep="\t",header=TRUE,as.is=TRUE)
  colnames(geno) <- gsub("-",".",colnames(geno))
  pheno$ID <- gsub("-",".",pheno$ID)
  rownames(pheno) <- pheno$ID
  netFile <- paste(path.package("netDx"),"extdata","CancerNets.txt",
  sep=getFileSep())
  cancerNets <- read.delim(netFile,sep="\t",header=T,as.is=T)
  message("* Excluding genes not present in interaction nets")
  noData <- setdiff(rownames(geno),rownames(cancerNets))
  if (length(noData)>0) {
  	message(paste(length(noData), 
  		" genes not present in cancer nets; excluding",sep=""))
  	geno <- geno[-which(rownames(geno) \%in\% noData),]
  }
  
  message("* Running label prop")
  require(doParallel)
  cl <- makeCluster(numCores)
  registerDoParallel(cl)
  prop_net <- smoothMutations_LabelProp(geno,cancerNets,cl,no_cores=numCores)
  stopCluster(cl)
  genoP <- thresholdSmoothedMutations(prop_net,geno,"OV_CancerNets")
}
