\name{toymodel}
\alias{toymodel}
\docType{data}
\title{
	Example model returned by a buildPredictor() call.
}
\description{
Output of buildPredictor() generated by a simple use-case of binary breast tumour classification using two genomic data sources.
BRCA data were downloaded using curatedTCGAData for mRNA and miRNA expression. buildPredictor()] was run by scoring features out of 2, with selected features passing 1 out of 2.
Tumours were labelled either "Luminal.A" or "other".
}
\usage{data(toymodel)}
\examples{
data(toymodel)
head(toymodel)
}
\keyword{datasets}
