% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEmapInput.R
\name{writeEMapInput}
\alias{writeEMapInput}
\title{write enrichment map for consensus nets}
\usage{
writeEMapInput(featScores, namedSets, netInfo, outPfx = "curr",
  pctPass = 0.7, maxScore = 10, trimFromName = c(".profile", "_cont"))
}
\arguments{
\item{featScores}{(data.frame) network scores across rounds of cross 
validation. Rows are networks and columns are network name followed by 
scores for cross-validation rounds. Output of getFeatureScores()}

\item{namedSets}{(list) list of nets and units (e.g.e pathway names and
genes). Should only contain units profiled in this dataset}

\item{netInfo}{(data.frame) Table of network name (netName) and type 
(netType). Type is used to assign shapes to nodes:
 clinical                                          clinical
      rna GUANOSINE_NUCLEOTIDES__I_DE_NOVO__I__BIOSYNTHESIS
      rna                              RETINOL_BIOSYNTHESIS}

\item{pctPass}{(numeric between 0 and 1) fraction of iterations that
a net's score must pass consCutoff for, to be included in the consensus
map}

\item{maxScore}{(integer) maximum possible score in one round of cross-
validation. e.g. for 10-fold cross-validation, maxScore=10.}

\item{trimFromName}{(char) strings to trim from name with sub()}

\item{consCutoff}{(integer) nets must pass this cutoff in at least
pctPass of the iterations to be written to file}
}
\value{
1) <outPfx>.gmt file - for enrichment map
2) <outPfx>_nodeAttr.txt (file) table with node properties, notably type,
pctPass
}
\description{
write enrichment map for consensus nets
}
