% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEmapInput_many.R
\name{writeEMapInput_many}
\alias{writeEMapInput_many}
\title{Wrapper to generate multiple EnrichmentMaps (perhaps one per class)}
\usage{
writeEMapInput_many(featScores, namedSets_valid, netTypes, outDir, ...)
}
\arguments{
\item{featScores}{(list) keys are classes, and values are data.frames of
network scores across cross-validation (output of getFeatScores()).}

\item{namedSets_valid}{(list) Grouped unit variables limited to the
units contained in the dataset. e.g. keys are pathways and values are
the genes measured in this dataset.
e.g.:
$`MISSPLICED_GSK3BETA_MUTANTS_STABILIZE_BETA-CATENIN`
[1] "PPP2R5E" "PPP2CB"  "APC"     "AXIN1"   "PPP2R1B" "PPP2R1A" "CSNK1A1"
[8] "PPP2R5D" "PPP2R5C" "PPP2R5B" "PPP2R5A" "PPP2CA"  "GSK3B"}

\item{netTypes}{(data.frame) "inputNets.txt" file
generated by NetDx. Dataframe has two columns, network type and
network  name. I.E:
 clinical                                          clinical
      rna GUANOSINE_NUCLEOTIDES__I_DE_NOVO__I__BIOSYNTHESIS
      rna                              RETINOL_BIOSYNTHESIS}

\item{outDir}{(char) path to output directory}

\item{...}{parameters for writeEMap()}
}
\description{
Wrapper to generate multiple EnrichmentMaps (perhaps one per class)
}
\examples{
data(featScores)

pathwayList <- readPathways(getExamplePathways())
pathwayList <- pathwayList[c(1:5)]

netInfoFile <- sprintf("\%s/extdata/example_output/inputNets.txt",
     path.package("netDx"))
netTypes <- read.delim(netInfoFile,sep="\\t",h=FALSE,as.is=TRUE)
outDir <- paste(tempdir(),"plots",sep="/")
if (!file.exists(outDir)) dir.create(outDir)
EMap_input <- writeEMapInput_many(featScores,pathwayList,
     netTypes,outDir=outDir)
}
