% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nets_writeSIF.R
\name{writeNetsSIF}
\alias{writeNetsSIF}
\title{write patient networks in Cytoscape's .sif format}
\usage{
writeNetsSIF(
  netPath,
  outFile = sprintf("\%s/out.sif", tempdir()),
  netSfx = "_cont.txt"
)
}
\arguments{
\item{netPath}{(char): vector of path to network files; file suffix
should be '_cont.txt' 
networks should be in format: A B 1
where A and B are nodes, and 1 indicates an edge between them}

\item{outFile}{(char) path to .sif file}

\item{netSfx}{(char) suffix for network file name}
}
\value{
No value. Side effect of writing all networks to \code{outFile}
}
\description{
write patient networks in Cytoscape's .sif format
}
\details{
Converts a set of binary interaction networks into Cytoscape's
sif format.
(http://wiki.cytoscape.org/Cytoscape_User_Manual/Network_Formats)
This utility permits visualization of feature selected networks.
}
\examples{
netDir <- system.file("extdata","example_nets",package="netDx")
netFiles <- sprintf('\%s/\%s', netDir, dir(netDir,pattern='txt$'))
writeNetsSIF(netFiles,'merged.sif',netSfx='.txt')
}
