% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeQueryBatchFile.R
\name{writeQueryBatchFile}
\alias{writeQueryBatchFile}
\title{Write batch.txt file required to create GeneMANIA database}
\usage{
writeQueryBatchFile(
  netDir,
  netList,
  outDir = tempdir(),
  idFile,
  orgName = "predictor",
  orgDesc = "my_predictor",
  orgAlias = "my_predictor",
  taxID = 1339
)
}
\arguments{
\item{netDir}{(char) path to dir with networks}

\item{netList}{(char) vector of network names}

\item{outDir}{(char) directory to write batch file}

\item{idFile}{(char) path to file with patient IDs}

\item{orgName}{(char) organism name. Don't change the default unless
you know what you are doing.}

\item{orgDesc}{(char) organism description. Similar to \code{orgName},
don't change the default}

\item{orgAlias}{(char) organism alias. Similar to \code{orgName}, don't
change the default.}

\item{taxID}{(integer) taxonomyID required for GeneMANIA . Similar to 
\code{orgName}, don't change the default.}
}
\value{
No value. Side effect of writing batch file to 
\code{<outDir>/batch.txt}.
}
\description{
Write batch.txt file required to create GeneMANIA database
}
\details{
This file is used to compile features into a single database
for feature selection.
}
\examples{
data(npheno)
netDir <- system.file("extdata","example_nets",package="netDx")
netList <- dir(netDir,pattern='txt$')
writeQueryBatchFile(netDir,netList, '~/tmp', npheno$ID)
}
