% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTopGroups.R
\name{writeTopGroups}
\alias{writeTopGroups}
\title{writes subset of gmt with feature-selected pathways}
\usage{
writeTopGroups(groupList, scores, cutoff, units = "*", outFile)
}
\arguments{
\item{groupList}{(list) names are input nets, values are member entities
e.g. pathways and genes}

\item{scores}{(data.frame) feature names (NAME) and score from predictor
building (SCORE)}

\item{cutoff}{(integer) only features with scores >= cutoff will be 
included}

\item{units}{(char) units that were present in the data. Not all units 
in groupList may be in the data. This function ensures that the units
written out are limited to those for which data were available
e.g. for a pathway, units are the genes that were present in the 
dataset. Genes that were not assayed would not be included in the output.
To include all genes (not recommended unless you know what you are doing)
, set to "*".}

\item{outFile}{(char) where output should be written.}
}
\value{
No value. Side effect of writing the gene-sets of features passing
`cutoff` to `outFile`.
}
\description{
writes subset of gmt with feature-selected pathways
}
