% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeWeightedNets.R
\name{writeWeightedNets}
\alias{writeWeightedNets}
\title{Write an integrated similarity network consisting of selected networks.}
\usage{
writeWeightedNets(patientIDs, netIDs, netDir, keepNets, filterEdgeWt = 0,
  aggNetFunc = "MAX", limitToTop = 50L, plotEdgeDensity = FALSE,
  verbose = FALSE)
}
\arguments{
\item{netDir}{(char) path to directory containing interaction networks.
Note that these are networks where the node IDs have been recoded by 
GeneMANIA (e.g. 1,2,3)}

\item{keepNets}{(char or data.frame) networks to include in integrated net
If data.frame must be in "NETWORK" column,other columns will be
ignored. Mainly included as convenience so pathway scores can passed
in table format
(NETWORK), and a multiplier constant for edges in that network (WEIGHT)}

\item{filterEdgeWt}{(numeric) keep edges with raw edge
weight strictly greater than this value. Note that "raw" refers to 
this filter being applied before the multiplier is applied.}

\item{aggNetFunc}{(char, one of: [MEAN|MAX]) Aggregate the network 
2) MEAN: average of weighted edges (raw x netDx score)
3) MAX: max of raw edge weight}

\item{limitToTop}{(integer) limit to top strongest connections. Set to
Inf to list all connections.}

\item{plotEdgeDensity}{(logical) plot density plot of edge weights, one
per input net. Used to troubleshoot problems introduced by specific nets.}

\item{verbose}{(logical) print messages if TRUE}

\item{patientID}{(data.frame) patient identifiers. Columns include
internally-generated identifiers (GM_ID) and user-provided identifiers 
(ID)}

\item{netInfo}{(data.frame) network metadata. Columns include
internal network name (NET_ID), user-provided name (NETWORK).
If a third optional column named "isBinary" is provided, and contains
binary values (i.e. 1 and 0), that indicates that the network contains
only binary weights and an alternate similarity computation (PropBinary)
will be used (see description).}
}
\value{
(list) 
1) filterEdgeWt (numeric) Value of filterEdgeWt parameter
2) aggNetFunc (char) Value of aggNetFunc parameter
3) limitToTop (integer) Value of limitToTop parameter
4) aggNet (matrix) Value of limitToTop parameter
File format is:
1) source patient (SOURCE)
2) target patient (TARGET)
3) network name (NET_NAME)
4) weight similarity for the network (WT_SIM)
}
\description{
Write an integrated similarity network consisting of selected networks.
}
