\name{xpr}
\alias{xpr}
\docType{data}
\title{
Example expression matrix
}
\description{
data.frame with gene expression for 727 genes (rows) and 40 patients (columns).
Data from TCGA breast cancer subtyping study.
}
\usage{data(xpr)}
\source{
The Cancer Genome Atlas. (2012). Nature 490:61-70.
}
\references{
The Cancer Genome Atlas. (2012). Nature 490:61-70.
}
\examples{
data(xpr)
head(xpr)
}
\keyword{datasets}
