% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPredictor_resampling_CNV.R
\name{buildPredictor_resampling_CNV}
\alias{buildPredictor_resampling_CNV}
\title{Build predictor with n-way resampling, for CNV-based nets}
\usage{
buildPredictor_resampling_CNV(pheno, p_GR, predClass, unitSet_GR, pctT = 0.7,
  numResamples = 3L, cliqueReps = 500L, nFoldCV = 10L, numCores = 1L,
  GMmemory = 4L, outDir = ".", overwrite = FALSE, seed_trainTest = 42L,
  seed_resampling = 103L, seed_CVqueries = 42L, ...)
}
\arguments{
\item{pheno}{(data.frame) patient IDs (ID) and class of interest (STATUS)}

\item{p_GR}{(GRanges) patient CNVs. Rows are patients, columns are unit
measures (e.g. genes)}

\item{predClass}{(char) class to build predictor for. Must be a value
in the STATUS column of the pheno param}

\item{unitSet_GR}{(list) unit groupings, each of which will be 
converted to its own patient similarity network.}

\item{pctT}{(numeric 0 to 1) fraction of samples for training}

\item{numResamples}{(integer) number of resamplings for cross validation}

\item{cliqueReps}{(integer) number of permutations to perform in
clique filtering}

\item{nFoldCV}{(integer) number of folds for cross-validation within each
resampling}

\item{numCores}{(integer) number of parallel processing jobs}

\item{GMmemory}{(integer) memory (Gb) for GeneMANIA to run.}

\item{outDir}{(char) path to output directory}

\item{overwrite}{(logical) Overwrite existing results in outDir? By 
default, no, return warning; but set to TRUE if outDir should be deleted
and populated with fresh results from this function.}

\item{seed_trainTest}{(integer) RNG seed for initial train/test split.
Set to NULL to use current state of RNG. Set to a constant for 
reproducibility}

\item{seed_resampling}{(integer) RNG seed for training assignment for
of the resamplings (splitTrainTest_partition() method)}

\item{seed_CVqueries}{(integer) RNG seed for queries in k-fold cross
validation (used by makeCVqueries() which is called by 
GM_runCV_featureSet())}

\item{...}{parameters for Nway_netSum()}
}
\value{
(list) Includes:
* pheno: (data.frame) phenotype matrix as used by predictor
* TT_STATUS: initial train/test split assignment
* netScores: (list) keys are classes, value is a data.frame with 
networks and scores following feature selection. Nets with score>0 are
shown.
* confmat: (matrix) confusion matrix with one row per cutoff
* predRanks: (list) GM rankings of test samples for each network cutoff
In addition, all the intervening work will be stored in <outDir>
}
\description{
Build predictor with n-way resampling, for CNV-based nets
}

