% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenet_enrichment_functions.R
\name{countIntType}
\alias{countIntType}
\title{Counts the number of (+,+) and (+,-) interactions in a single network}
\usage{
countIntType(inFile, plusID, minusID)
}
\arguments{
\item{inFile}{(char) path to interaction networks}

\item{plusID}{(char) vector of + nodes}

\item{minusID}{(char) vector of - nodes}
}
\value{
(numeric of length 2) Number of (+,+) interactions, and 
non-(+,+) interactions
(i.e. (+,-) and (-,-) interactions)
}
\description{
Counts the number of (+,+) and (+,-) interactions in a single network
}
\examples{
d <- tempdir()
# write PSN
m1 <- matrix(c("P1","P1","P2","P2","P3","P4",1,1,1),byrow=FALSE,ncol=3)
write.table(m1,file=sprintf("\%s/net1.txt",d),sep="\\t",
col.names=FALSE,row.names=FALSE,quote=FALSE)

countIntType(sprintf("\%s/net1.txt",d),c("P1","P2","P3"),
c("P4","P5"))
}
