% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countIntType_batch.R
\name{countIntType_batch}
\alias{countIntType_batch}
\title{Counts number of (+,+) and (+,-) interactions in a set of networks}
\usage{
countIntType_batch(inFiles, plusID, minusID, tmpDir = "/tmp",
  enrType = "binary")
}
\arguments{
\item{inFiles}{(char) path to interaction networks to process}

\item{plusID}{(char) IDs of + nodes}

\item{minusID}{(char) IDs of - nodes}

\item{tmpDir}{(char) path to dir where temporary files can be stored}

\item{enrType}{(char) see getEnr.R}
}
\value{
(matrix) two columns, one row per network 
If \code{enrType="binary"}, number of (+,+) and other interactions
Otherwise if \code{enrType="corr"} mean edge weight of (+,+) edges and
of other edges
}
\description{
Counts number of (+,+) and (+,-) interactions in a set of networks
}

