% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsenet_enrichment_functions.R
\name{countIntType_batch}
\alias{countIntType_batch}
\title{Counts number of (+,+) and (+,-) interactions in a set of networks}
\usage{
countIntType_batch(
  inFiles,
  plusID,
  minusID,
  tmpDir = tempdir(),
  enrType = "binary",
  numCores = 1L
)
}
\arguments{
\item{inFiles}{(char) path to interaction networks to process}

\item{plusID}{(char) IDs of + nodes}

\item{minusID}{(char) IDs of - nodes}

\item{tmpDir}{(char) path to dir where temporary files can be stored}

\item{enrType}{(char) see getEnr.R}

\item{numCores}{(integer) number of cores for parallel processing}
}
\value{
(matrix) two columns, one row per network 
If \code{enrType="binary"}, number of (+,+) and other interactions
Otherwise if \code{enrType="corr"} mean edge weight of (+,+) edges and
of other edges
}
\description{
Counts number of (+,+) and (+,-) interactions in a set of networks
}
\examples{
d <- tempdir()
# write PSN
m1 <- matrix(c("P1","P1","P2","P2","P3","P4",1,1,1),byrow=FALSE,ncol=3)
write.table(m1,file=paste(d,"net1.txt",sep=.Platform$file.sep),sep="\t",
col.names=FALSE,row.names=FALSE,quote=FALSE)
m2 <- matrix(c("P3","P4",1),nrow=1)
write.table(m2,file=paste(d,"net2.txt",sep=.Platform$file.sep),sep="\t",
col.names=FALSE,row.names=FALSE,quote=FALSE)

countIntType_batch(paste(d,c("net1.txt","net2.txt"),sep=.Platform$file.sep),
	c("P1","P2","P3"),c("P4","P5"))
}
