% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataList2list.R
\name{dataList2List}
\alias{dataList2List}
\title{Convert MultiAssayExperiment object to list and data.frame}
\usage{
dataList2List(dat)
}
\arguments{
\item{dat}{(MultiAssayExperiment)}
}
\value{
(list) Keys are:
1) assays: list of matrices, each corresponding to data from a particular
layer
2) pheno: (data.frame) sample metadata
}
\description{
Convert MultiAssayExperiment object to list and data.frame
}
\details{
Used by internal routines in netDx
}
