% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSubset.R
\name{doSubsetSelection}
\alias{doSubsetSelection}
\title{Feature subset selection driver function}
\usage{
doSubsetSelection(methodName = "greedy.forward", incNets, queryPool, GM_db,
  pheno_DF, predictClass, outDir, num2return, numCores = 1L,
  scoreFile = NULL)
}
\arguments{
\item{methodName}{(char) subset selection strategy. Options are:
1) greedy.forward : forward selection
2) greedy.backward : backward elimination
3) hill.climb: hill climbinb}

\item{incNets}{(char) vector of networks that are the basis of subset
selection}

\item{queryPool}{(char) See GM_run_featureSet, trainID param}

\item{GM_db}{(char) path to GeneMANIA generic database with 
training population. This is the same as the \code{dbDir} value for
the output of GM_createDB.R}

\item{pheno_DF}{(data.frame) pheno matrix; ID and STATUS}

\item{predictClass}{(char) class label for predictor}

\item{outDir}{(char) directory to store query file and GM results}

\item{num2return}{(integer) number of training samples in total}

\item{numCores}{(integer) num cores for parallel processing. Used by
\code{GM_runCV_featureSet()} during the cross-validation for each 
feature subset}

\item{scoreFile}{(char) path to file where error is printed over 
iterations. useful for debugging. If NULL doesn't print.}
}
\value{
(char) vector of selected subset of networks
}
\description{
Feature subset selection driver function
}
\details{
Wrapper to run feature subselection methods from the FSelector 
package
}
\examples{
GM_db <- sprintf("\%s/extdata/GM_db",path.package("netDx"))
data(MB_pheno)
nets <- c("DKK2_cont","EMX2_cont","TNC_cont","WIF1_cont")
doSubsetSelection(methodName="greedy.backward",incNets=nets, 
queryPool=MB.pheno$ID[which(MB.pheno$STATUS\%in\% "WNT")],
GM_db=GM_db,pheno_DF=MB.pheno, outDir=".",num2return=103L)
}
