% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileCache.R
\name{fetchPathwayDefinitions}
\alias{fetchPathwayDefinitions}
\title{fetch pathway definitions from downloads.baderlab.org}
\usage{
fetchPathwayDefinitions(month = NULL, year = NULL, day = 1, verbose = FALSE)
}
\arguments{
\item{month}{(numeric or char) month of pathway definition file. Can be
numeric or text (e.g. "January","April"). If NULL, fails.}

\item{year}{(numeric) year of pathway definition file. Must be in
yyyy format (e.g. 2018). If NULL, fails.}

\item{day}{(integer)}

\item{verbose}{(logical) print messages}
}
\value{
(char) Path to local cached copy of GMT file
or initial download is required
}
\description{
fetch pathway definitions from downloads.baderlab.org
}
\details{
Fetches genesets compiled from multiple curated pathway
databases. Downloaded from: https://download.baderlab.org/EM_Genesets/
The file contains pathways from HumanCyc, NetPath, Reactome, NCI
Curated Pathways and mSigDB.
For details see Merico D, Isserlin R, Stueker O, Emili A and GD Bader.
(2010). PLoS One. 5(11):e13984.
}
\examples{
fetchPathwayDefinitions("October",2020)
fetchPathwayDefinitions("January",2018)
fetchPathwayDefinitions(month=10,year=2020)
}
