% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatureScores.R
\name{getFeatureScores}
\alias{getFeatureScores}
\title{Compile network scores into a matrix}
\usage{
getFeatureScores(inDir, predClasses, getFullCons = TRUE)
}
\arguments{
\item{inDir}{(char/list) directory containing directories with all split info
or list of all CV score files.
if inDir is a single directory then the expected format for CV score files is
<inDir>/rngX/predClassX/GM_results/predClassX_pathway_CV_score.txt"
if inDir is a list, it should have one key per class. The value should be the
corresponding set of filenames for pathway_CV_score.txt}

\item{predClasses}{(char) possible STATUS for patients}

\item{getFullCons}{(logical) if TRUE, does not remove rows with NA.
Recommended only when the number of input features is extensively 
pruned by first-pass feature selection.}
}
\value{
(list) one key per patient class. Value is matrix of network
scores across all train/test splits. Each score is the output of
the inner fold of CV.
}
\description{
Compile network scores into a matrix
}
\details{
Given network scores over a set of trian/test splits, compiles these
into a matrix for downstream analysis. See the section on "Output Files"
}
\examples{
inDir <- sprintf("\%s/extdata/example_output",path.package("netDx"))
netScores <- getFeatureScores(inDir, predClasses = c("SURVIVEYES","SURVIVENO"))
}
