% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getPSN}
\alias{getPSN}
\title{get the integrated patient similarity network made of selected features}
\usage{
getPSN(
  dat,
  groupList,
  makeNetFunc = NULL,
  sims = NULL,
  selectedFeatures,
  plotCytoscape = FALSE,
  aggFun = "MEAN",
  prune_pctX = 0.3,
  prune_useTop = TRUE,
  numCores = 1L,
  calcShortestPath = FALSE
)
}
\arguments{
\item{dat}{(MultiAssayExperiment) input data}

\item{groupList}{(list) feature groups, identical to groupList provided for buildPredictor()}

\item{makeNetFunc}{(function) Function used to create patient similarity networks. Identical to 
makeNets provided to buildPredictor()}

\item{sims}{(list) rules for creating PSN. Preferred over makeNetFunc. See buildPredictor() 
for details.}

\item{selectedFeatures}{(list) selected features for each class (key of list). This object is returned as
part of a call to getResults(), after running buildPredictor().}

\item{plotCytoscape}{(logical) If TRUE, plots network in Cytoscape.
Requires Cytoscape software to be installed and running on the computer
when the function call is being made.}

\item{aggFun}{(char) function to aggregate edges from different PSN (e.g. mean)}

\item{prune_pctX}{(numeric between 0 and 1) fraction of most/least 
edges to keep when pruning the integrated PSN for visualization.
Must be used in conjunction with useTop=TRUE/FALSE
e.g. Setting pctX=0.2 and useTop=TRUE will keep 20\% top edges}

\item{prune_useTop}{(logical) when pruning integrated PSN for visualization,
determines whether to keep strongest edges (useTop=TRUE) or weakest edges
(useTop=FALSE)}

\item{numCores}{(integer) number of cores for parallel processing}

\item{calcShortestPath}{(logical) if TRUE, computes weighted shortest path
Unless you plan to analyse these separately from looking at the shortest 
path violin plots or integrated PSN in Cytoscape, probably good to set to 
FALSE.}
}
\value{
(list) information about the integrated network
similarity network
2) patientDistNetwork_pruned (matrix) the network plotted in
Cytoscape. Also note that this is a dissimilarity network, 
so that more similar nodes have smaller edge weights
3) colLegend (data.frame): legend for the patient network
plotted in Cytoscape. Columns are node labels (STATUS) and
colours (colour)
6) outDir (char) value of outDir parameter
}
\description{
get the integrated patient similarity network made of selected features
}
\details{
An integrated patient similarity network can be built using combined
top features for each patient class. Such a network is created by taking the union of selected features for
all patient labels, and aggregating pairwise edges for all of them using a user-specified function (aggFun).
The network is then pruned prior to visualization, using a user-specified fraction of strongest edges
(prune_pctX, prune_useTop). In addition, the user may quantify the distance between patients of the 
same class, relative to those of other classes, using Dijkstra distance (calcShortestPath flag).
}
