% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getResults}
\alias{getResults}
\title{prints performance, returns list of output.}
\usage{
getResults(
  dat,
  res,
  featureSelCutoff = 1L,
  featureSelPct = 0,
  pathwayList = NULL
)
}
\arguments{
\item{dat}{(MultiAssayExperiment) input data}

\item{res}{(list) output of buildPredictor() function}

\item{featureSelCutoff}{(integer) cutoff score for feature selection.
A feature must have minimum of this score for specified fraction of splits 
(see featureSelPct) to pass.}

\item{featureSelPct}{(numeric between 0 and 1) cutoff percent for feature selection.
A feature must have minimum score of featureSelCutoff for featureSelPct of 
train/test splits to pass.}
}
\value{
list of results.
- selectedFeatures (list of character vectors): list, one per class
- performance (list of mixed datatypes) accuracy, confusion matrix, AUROC
Side effect of plotting ROC curve if binary classifier
}
\description{
prints performance, returns list of output.
}
