% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getResults}
\alias{getResults}
\title{Compiles performance and selected features for a trained model.}
\usage{
getResults(
  res,
  status,
  featureSelCutoff = 1L,
  featureSelPct = 0,
  drawPerformancePlot = TRUE
)
}
\arguments{
\item{res}{(list) output of buildPredictor() function}

\item{status}{(character) unique patient labels used by the classifier, found in colData()$STATUS}

\item{featureSelCutoff}{(integer) cutoff score for feature selection.
A feature must have minimum of this score for specified fraction of splits 
(see featureSelPct) to pass.}

\item{featureSelPct}{(numeric between 0 and 1) cutoff percent for feature selection.
A feature must have minimum score of featureSelCutoff for featureSelPct of 
train/test splits to pass.}

\item{drawPerformancePlot}{(logical) if TRUE, draws AUROC and AUPR plots. Set to FALSE to suppress graphical output.}
}
\value{
list of results.
- selectedFeatures (list of character vectors): list, one per class
- performance (list of mixed datatypes) including mean accuracy (meanAccuracy), 
split-level accuracy (splitAccuracy), split-level AUROC (auroc),
split-level AUPR (splitAUR)
Side effect of plotting ROC curve if binary classifier
}
\description{
Compiles performance and selected features for a trained model.
}
\details{
This function is run after training a model using buildPredictor(). 
It takes patient input data, model output, and returns performance and selected features.
}
\examples{
data(toymodel) # load example results from binary breast classification
patlabels <- names(toymodel$Split1$featureSelected)
getResults(toymodel,patlabels,2,0.5)

}
