% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_predClass_once.R
\name{GM_predClass_once}
\alias{GM_predClass_once}
\title{wrapper to build PSN + GM database + run query + get rankings for N-way}
\usage{
GM_predClass_once(pheno, pdat, predClass, unitSets, p_GR = NULL,
  unitSet_GR = NULL, patNets, outDir, numCores = 1L)
}
\arguments{
\item{pheno}{(data.frame) patient metadata. ID, class labels (STATUS),
and whether train/test status (TT_STATUS): a patient is query if "TRAIN",
else is "TEST"}

\item{pdat}{(data.frame) patient data. Rows are unit variables, columns 
are patients. Rows and columns must be named.}

\item{predClass}{(char) class of interest to predict.}

\item{unitSets}{(list) groupings of unit variables; each grouping will
be converted into its own PSN.}

\item{p_GR}{(GRanges) GRanges of patient CNVs. Has ID column in
metadata, containing patient IDs}

\item{unitSet_GR}{(list) sets of GRanges to group CNVs (e.g.
could have one GRanges per pathway, corresponding to regions in that 
pathway}

\item{patNets}{(list of chars) for each class (key), subset of "unitSets"
for which patient nets must be created. e.g. feature-selected networks
for the corresponding patient subtype.}

\item{outDir}{(char) directory in which results must be written}

\item{numCores}{(integer) number of cores for parallel processing}
}
\value{
(list) 1. predRes: patient rankings for each class; 
2. predClass: predicted patient labels 
3. perfStats (list): model performance stats: tp,fp,tn,fn,accuracy,ppv
}
\description{
wrapper to build PSN + GM database + run query + get rankings for N-way
}
\details{
This set of steps could be used for final model performance
once feature selection is complete. The steps are { build networks, 
build GM database, run query } once per class. Finally rank patients
based on best class. Currently supports only full matrix data and only
works for two-way classification (predClass/"other"). 
Extension to N-way classification
probably doesn't require much additional work
}
