% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{makeInputForEnrichmentMap}
\alias{makeInputForEnrichmentMap}
\title{Wrapper to create input files for Enrichment Map}
\usage{
makeInputForEnrichmentMap(
  model,
  results,
  pathwayList,
  EMapMinScore = 0L,
  EMapMaxScore = 1L,
  EMapPctPass = 0.5,
  outDir
)
}
\arguments{
\item{model}{(list) Output of training model, generated by running buildPredictor()}

\item{results}{(list) Model results. output of getResults()}

\item{pathwayList}{(list) output of readPathwayFile() used to make pathway-level feat ures for predictor}

\item{EMapMinScore}{(integer) minimum score for Enrichment Map}

\item{EMapMaxScore}{(integer) maximum score for Enrichment Map}

\item{EMapPctPass}{(numeric between 0 and 1) percent of splits for which feature must have score in range
[EMapMinScore,EMapMaxScore] to be included for EnrichmentMap visualization}

\item{outDir}{(char) directory where files should be written}
}
\value{

}
\description{
Wrapper to create input files for Enrichment Map
}
\details{
An Enrichment Map is a network-based visualization of top-scoring pathway features
and themes. It is generated in Cytoscape. This script generates the input files needed
for Cytoscape to create an Enrichment Map visualization.
}
