% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmap.R
\name{plotEmap}
\alias{plotEmap}
\title{Using GMT files, generate enrichment maps and apply Auto annotate app to them}
\usage{
plotEmap(gmtFile, nodeAttrFile, netName = "generic", outDir, minScore = 1,
  maxScore = 10, cytoscapeBaseURL = "http://localhost:1234/v1",
  verbose = FALSE)
}
\arguments{
\item{gmtFile}{(char) file path to GMT file (generated by writeEMap_file()).
NOTE: This needs to be the absolute path name}

\item{nodeAttrFile}{(list) file path to nodeAttr.txt file (generated by 
writeEMap_file())}

\item{netName}{(char) name for network in Cytoscape. Using the patient 
class name is a good idea. (e.g. SURVIVE_YES and SURVIVE_NO).}

\item{outDir}{(char) path to directory where file should be stored.}

\item{cytoscapeBaseURL}{(char) URL for CyREST connection to the active 
Cytoscape session. Only change this if you know what you're doing.}

\item{verbose}{(logical) print messages}
}
\description{
Using GMT files, generate enrichment maps and apply Auto annotate app to them
}
\examples{
plotEmap(gmt_file, node_attrFile, netName="HighRisk")
}
