% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmap.R
\name{plotEmap}
\alias{plotEmap}
\title{Using GMT files, generate enrichment maps and apply Auto annotate app to them}
\usage{
plotEmap(gmtFile, nodeAttrFile, netName = "generic",
  p_value_cutoff = "0.005", q_value_cutoff = "0.01")
}
\arguments{
\item{gmtFile}{(char) file path to GMT file (generated by writeEMap_file()).
NOTE: This needs to be the absolute path name}

\item{nodeAttrFile}{(list) file path to nodeAttr.txt file (generated by 
writeEMap_file())}

\item{p_value_cutoff}{(str) p-value cutoff input for enrichment map. All gene
sets with a p-value with the specified threshold or below are included in 
the map.}

\item{q_value_cutoff}{(str) FDR q-value cutoff input for enrichment map. 
All gene sets with FDR q-value with the specified threshold or below are 
included in the map.}
}
\description{
Using GMT files, generate enrichment maps and apply Auto annotate app to them
}
\examples{
plotEmap(gmt_file, node_attrFile, netName="HighRisk",
	p_value_cutoff = "0.005", q_value_cutoff = "0.01")
}
