% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmap.R
\name{plotEmap}
\alias{plotEmap}
\title{Create EnrichmentMap in Cytoscape to visualize predictive pathways}
\usage{
plotEmap(gmtFile, nodeAttrFile, netName = "generic", outDir,
  minScore = 1, maxScore = 10, colorScheme = "cont_heatmap",
  imageFormat = "png", verbose = FALSE, createStyle = TRUE)
}
\arguments{
\item{gmtFile}{(char) file path to GMT file (generated by writeEMap_file()).
NOTE: This needs to be the absolute path name}

\item{nodeAttrFile}{(list) file path to nodeAttr.txt file (generated by
writeEMap_file())}

\item{netName}{(char) name for network in Cytoscape. Using the patient
class name is a good idea. (e.g. SURVIVE_YES and SURVIVE_NO).}

\item{outDir}{(char) path to directory where file should be stored.}

\item{minScore}{(int) minimum score of node to show}

\item{maxScore}{(int) maximum score of node to show}

\item{colorScheme}{(char) colour scheme for nodes. "cont_heatmap" 
sets a discrete map ranging from yellow to red for increasing scores. 
"netDx_ms" is the colour scheme used in the netDx methods paper.
This map is (<=6: white; 7-9: orange; 10: red)}

\item{imageFormat}{(char) one of PNG, PDF, SVG, or JPEG}

\item{verbose}{(logical) print messages}

\item{createStyle}{(logical) if generating more than one EMap, set to 
TRUE for first one and to FALSE for subsequent. Due to limitation in 
current version of RCy3}
}
\value{
Filename of image to which EnrichmentMap is exported. Also side
effect of plotting the EnrichmentMap in an open session of Cytoscape.
}
\description{
Create a network where nodes are predictive pathways passing
certain cutoff and edges indicate similarity in gene-sets. Pathways are then
clustered to identify themes of predictive pathways. Generates one such
network for each patient label.
}
\examples{
#refer to writeEMapInput_many.R for working writeEMapInput_many() example
data(featScores)
pathFile <- sprintf("\%s/extdata/Human_160124_AllPathways.gmt",
          path.package("netDx.examples"))
pathwayList <- readPathways(pathFile)
pathwayList <- pathwayList[c(1:5)]
netInfoFile <- sprintf("\%s/extdata/KIRC_output/inputNets.txt",
     path.package("netDx.examples"))
netTypes <- read.delim(netInfoFile,sep="\\t",h=FALSE,as.is=TRUE)
outDir <- paste(tempdir(),"plots",sep="/")
if (!file.exists(outDir)) dir.create(outDir)
EMap_input <- writeEMapInput_many(featScores,pathwayList,
     netTypes,outDir=outDir)
outDir <- paste(getwd(),"plots",sep="/")
if (!file.exists(outDir)) dir.create(outDir)
gmtFile <- EMap_input[[1]][1]
nodeAttrFile <- EMap_input[[1]][2]

# not run because requires Cytoscape to be installed and open
# plotEmap(gmtFile = gmtFile, nodeAttrFile = nodeAttrFile, netName="HighRisk",
# outDir=outDir)
}
