% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmap.R
\name{plotEmap}
\alias{plotEmap}
\title{Using GMT files, generate enrichment maps and apply Auto annotate app to them}
\usage{
plotEmap(gmtFile, nodeAttrFile, netName = "generic", outDir,
  minScore = 1, maxScore = 10, colorScheme = "cont_heatmap",
  imageFormat = "png", verbose = FALSE, createStyle = TRUE)
}
\arguments{
\item{gmtFile}{(char) file path to GMT file (generated by writeEMap_file()).
NOTE: This needs to be the absolute path name}

\item{nodeAttrFile}{(list) file path to nodeAttr.txt file (generated by
writeEMap_file())}

\item{netName}{(char) name for network in Cytoscape. Using the patient
class name is a good idea. (e.g. SURVIVE_YES and SURVIVE_NO).}

\item{outDir}{(char) path to directory where file should be stored.}

\item{minScore}{(int) minimum score of node to show}

\item{maxScore}{(int) maximum score of node to show}

\item{colorScheme}{(char) colour scheme for nodes. "cont_heatmap" 
sets a discrete map ranging from yellow to red for increasing scores. 
"netDx_ms" is the colour scheme used in the netDx methods paper.
This map is (<=6: white; 7-9: orange; 10: red)}

\item{verbose}{(logical) print messages}

\item{createStyle}{(logical) if generating more than one EMap, set to 
TRUE for first one and to FALSE for subsequent. Due to limitation in 
current version of RCy3}

\item{imageFormat(char)}{one of PNG, PDF, SVG, or JPEG}
}
\description{
Using GMT files, generate enrichment maps and apply Auto annotate app to them
}
\examples{
#refer to writeEMapInput_many.R for working writeEMapInput_many() example
EMap_input <- writeEMapInput_many(featScores,pathwayList,
     netInfo,outDir=outDir)
outDir <- paste(getwd(),"plots",sep="/")
if (!file.exists(outDir)) dir.create(outDir)
gmtFile <- EMap_input[[1]][1]
nodeAttrFile <- EMap_input[[1]][2]
plotEmap(gmtFile = gmtFile, nodeAttrFile = nodeAttrFile, netName="HighRisk",
outDir=outDir)
}
