% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_diagnostics.R
\name{plotNetEdgeDistribution}
\alias{plotNetEdgeDistribution}
\title{plot distribution of PSN edge weights}
\usage{
plotNetEdgeDistribution(netDir, setName = "netDir",
  randomSample = NULL, grepFiles = "cont.txt",
  stripFromFileName = "_cont.txt", numPatients = NULL,
  verbose = TRUE)
}
\arguments{
\item{netDir}{(char) path to directory of networks. Networks must be
in SIF format; i.e. with three columns: source, target, edge weight.}

\item{setName}{(char) used for output file name (<setName>_edgeDistr.pdf)}

\item{randomSample}{(integer) if set to NULL. Plots for all networks.
If set to an integer plots the number of networks requested, to the max
number of networks available}

\item{grepFiles}{(char) searches for files with the provided string.}

\item{stripFromFileName}{(char) vector of char patterns to remove from file name}

\item{verbose}{(logical) print messages}
}
\description{
plot distribution of PSN edge weights
}
\details{
Used for comparing the properties of PSN for different similarity
metrics
}
