% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPerf.R
\name{plotPerf}
\alias{plotPerf}
\title{Plots various measures of predictor performance for binary classifiers}
\usage{
plotPerf(inDir, predClasses)
}
\arguments{
\item{inDir}{(char) path to predictionResults.txt files.
if inDir is a single char and not a vector of prediction files, the expected
directory structure is <inDir>/rng<X>/predictionResults.txt".
Otherwise inDir is assumed
to be a vector, each with absolute paths to predictionResults.txt}

\item{predClasses}{(char) vector of class names.}
}
\value{
(list) each key corresponds to an input file in inDir.
Value is a list with:
1) stats: "stats" component of perfCalc
2) rocCurve: ROCR performance object for ROC curve
3) prCurve: ROCR performance object for PR curve
4) auroc: Area under ROC curve
5) aupr: Area under PR curve
6) accuracy: Accuracy

Side effect of plotting in a 2x2 format:
1) mean+/-SEM AUROC
2) mean+/-SEM AUPR
3) ROC curve for all runs plus average
4) PR curve for all runs plus average
}
\description{
Plots various measures of predictor performance for binary classifiers
}
\details{
Plots individual and average ROC/PR curves. mean+/-SEM performance for a predictor run using nested
cross-validation or a similar repeated design.
predictionResults.txt contains a (data.frame)
}
\examples{
inDir <- sprintf("\%s/extdata/example_output",
	path.package("netDx"))
plotPerf(inDir, predClasses = c("LumA","notLumA"))
}
