% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPerf_multi.R
\name{plotPerf_multi}
\alias{plotPerf_multi}
\title{Plots a set of ROC/PR curves with average.}
\usage{
plotPerf_multi(inList, plotTitle = "performance", plotType = "ROC",
  xlab = "TPR", ylab = "FPR", meanCol = "darkblue", xlim = c(0, 1),
  ylim = c(0, 1))
}
\arguments{
\item{inList}{(list) ROCR::performance objects, one per iteration}

\item{plotTitle}{(numeric) plot title}

\item{plotType}{(char) one of ROC | PR | custom. Affects x/y labels}

\item{xlab}{(char) x-axis label}

\item{ylab}{(char) y-axis label}

\item{meanCol}{(char) colour for mean trendline}

\item{xlim}{(numeric) min/max extent for x-axis}

\item{ylim}{(numeric) min/max extent for y-axis}
}
\value{
No value. Side effect of plotting ROC and PR curves
}
\description{
Plots a set of ROC/PR curves with average.
}
\details{
Plots average curves with individual curves imposed.
}
\examples{
inDir <- sprintf("\%s/extdata/KIRC_output", path.package("netDx.examples"))
all_rng <- list.files(path = inDir, pattern = "rng.")
fList <- sprintf("\%s/\%s/predictionResults.txt", inDir,all_rng)
dat <- read.delim(fList[1],sep="\\t",h=T,as.is=T)
predClasses <- c('SURVIVEYES', 'SURVIVENO')
pred_col1 <- sprintf("\%s_SCORE",predClasses[1])
pred_col2 <- sprintf("\%s_SCORE",predClasses[2])
idx1 <- which(colnames(dat) == pred_col1)
idx2 <- which(colnames(dat) == pred_col2)
pred <- ROCR::prediction(dat[,idx1]-dat[,idx2], dat$STATUS==predClasses[1])
curRoc <- ROCR::performance(pred,"tpr","fpr")
plotPerf_multi(list(curRoc=curRoc),"ROC")
}
