% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPerf_multi.R
\name{plotPerf_multi}
\alias{plotPerf_multi}
\title{Plots a set of ROC/PR curves with average.}
\usage{
plotPerf_multi(
  inList,
  plotTitle = "performance",
  plotType = "ROC",
  xlab = "TPR",
  ylab = "FPR",
  meanCol = "darkblue",
  xlim = c(0, 1),
  ylim = c(0, 1)
)
}
\arguments{
\item{inList}{(list) ROCR::performance objects, one per iteration}

\item{plotTitle}{(numeric) plot title}

\item{plotType}{(char) one of ROC | PR | custom. Affects x/y labels}

\item{xlab}{(char) x-axis label}

\item{ylab}{(char) y-axis label}

\item{meanCol}{(char) colour for mean trendline}

\item{xlim}{(numeric) min/max extent for x-axis}

\item{ylim}{(numeric) min/max extent for y-axis}
}
\value{
No value. Side effect of plotting ROC and PR curves
}
\description{
Plots a set of ROC/PR curves with average.
}
\details{
Plots average curves with individual curves imposed.
}
\examples{
inDir <- system.file("extdata","example_output",package="netDx")
all_rng <- list.files(path = inDir, pattern = 'rng.')
fList <- paste(inDir,all_rng,'predictionResults.txt',sep=getFileSep())
rocList <- list()
for (k in seq_len(length(fList))) {
  dat <- read.delim(fList[1],sep='\t',header=TRUE,as.is=TRUE)
  predClasses <- c('LumA', 'notLumA')
  pred_col1 <- sprintf('\%s_SCORE',predClasses[1])
  pred_col2 <- sprintf('\%s_SCORE',predClasses[2])
  idx1 <- which(colnames(dat) == pred_col1)
  idx2 <- which(colnames(dat) == pred_col2)
 pred <- ROCR::prediction(dat[,idx1]-dat[,idx2], 
		dat$STATUS==predClasses[1])
 rocList[[k]] <- ROCR::performance(pred,'tpr','fpr')
}
plotPerf_multi(rocList,'ROC')
}
