% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictPatientLabels.R
\name{predictPatientLabels}
\alias{predictPatientLabels}
\title{assign patient class when ranked by multiple GM predictors}
\usage{
predictPatientLabels(resSet)
}
\arguments{
\item{resSet}{(list) output of getPatientRankings, each key for a different
predictor. names(resSet) contain predictor label}
}
\value{
data.frame: ID, GM_score, PRED_CLASS
}
\description{
assign patient class when ranked by multiple GM predictors
}
\examples{
data(predRes); predClass <- predictPatientLabels(predRes)
}
