% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{psn__builtIn}
\alias{psn__builtIn}
\title{make PSN for built-in similarity functions}
\usage{
psn__builtIn(settings, verbose, ...)
}
\arguments{
\item{settings}{(list) from makeNetFunc}

\item{verbose}{(logical) print messages}

\item{...}{parameters for makePSN_NamedMatrix()}
}
\value{
(char) names of networks created. Side effect of network creation.
}
\description{
make PSN for built-in similarity functions
}
