% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{psn__corr}
\alias{psn__corr}
\title{wrapper for PSNs using Pearson correlation}
\usage{
psn__corr(settings, verbose, ...)
}
\arguments{
\item{settings}{(list) from makeNetFunc}
}
\description{
wrapper for PSNs using Pearson correlation
}
