% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{psn__custom}
\alias{psn__custom}
\title{make PSN for custom similarity functions}
\usage{
psn__custom(settings, fn, verbose, ...)
}
\arguments{
\item{settings}{(list) from makeNetFunc}

\item{fn}{(function) custom similarity function}

\item{verbose}{(logical) print messages}

\item{...}{parameters for makePSN_NamedMatrix()}
}
\value{
(char) names of networks created. Side effect of network creation.
}
\description{
make PSN for custom similarity functions
}
