% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_predTest_CNV.R
\name{resampling_predTest_CNV}
\alias{resampling_predTest_CNV}
\title{evaluate test performance for CNV-based predictor}
\usage{
resampling_predTest_CNV(testPheno, test_p_GR, selFeature_GR, predClass,
  cliqueNetGenes = NULL, verbose = TRUE)
}
\arguments{
\item{testPheno}{(data.frame) sample metadata table with ID and STATUS}

\item{test_p_GR}{(GRanges) genomic events in patients. Must have ID column
that maps an event to a patient, and LOCUS_NAMES with comma-separated names
of units (e.g. genes) that the ranges overlap}

\item{selFeature_GR}{(GRanges) genomic ranges for selected features
NOTE: These are the features over which patients will be evaluated so 
limit to nets that pass feature selection. An example is the set of genes
belonging to feature-selected pathways}

\item{predClass}{(char) label for positve class}

\item{cliqueNetGenes}{(GRanges) optional. If provided, universe is 
limited to patients overlapping ranges in this set. e.g. the set of genes
in all pathways that pass feature selection}

\item{verbose}{(logical) print messages}
}
\value{
(list) 
1) pos (char): vector of patients called +
2) neg (char): vector of patients called -
3-6) tp,tn,acc,ppv (numeric): num true pos, num true neg, accuracy and
positive predictive value
}
\description{
evaluate test performance for CNV-based predictor
}
