% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFeatureSelection.R
\name{runFeatureSelection}
\alias{runFeatureSelection}
\title{Run GeneMANIA cross-validation with a provided subset of networks}
\usage{
runFeatureSelection(
  trainID_pred,
  outDir,
  dbPath,
  numTrainSamps = NULL,
  incNets = "all",
  orgName = "predictor",
  fileSfx = "CV",
  verbose = FALSE,
  numCores = 2L,
  JavaMemory = 6L,
  verbose_runQuery = FALSE,
  ...
)
}
\arguments{
\item{trainID_pred}{(char) vector with universe of predictor class
patients (ie all that can possibly be included in the query file}

\item{outDir}{(char) directory to store query file and GM results}

\item{dbPath}{(char) path to GeneMANIA generic database with
training population}

\item{numTrainSamps}{(integer) number of training samples in total
leave blank to use 5 training samples in order to save memory}

\item{incNets}{(char) vector of networks to include in this analysis
(features/pathway names). Useful for subset-based feature selection}

\item{orgName}{(char) organism name for GeneMANIA generic database.
The default value will likely never need to be changed.}

\item{fileSfx}{(char) file suffix}

\item{verbose}{(logical) print messages}

\item{numCores}{(logical) num parallel threads for cross-validation}

\item{JavaMemory}{(integer) memory for GeneMANIA run, in Gb.}

\item{verbose_runQuery}{(logical) print messages for runQuery()}

\item{...}{args for \code{makeQueries()}}
}
\value{
No value. Side effect of generating feature scores.
}
\description{
Run GeneMANIA cross-validation with a provided subset of networks
}
\details{
Creates query files, runs GM for 10-fold cross validation.
}
\examples{
data(MB.pheno)
dbPath <- system.file("extdata","dbPath",package="netDx")
runFeatureSelection(MB.pheno$ID[which(MB.pheno$STATUS\%in\% 'WNT')],
		'~/tmp',dbPath,103L)
}
