% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQuery.R
\name{runQuery}
\alias{runQuery}
\title{Run a query}
\usage{
runQuery(dbPath, queryFiles, resDir, verbose = TRUE, JavaMemory = 6L,
  numCores = 1L, debugMode = FALSE)
}
\arguments{
\item{dbPath}{(char) path to directory with GeneMANIA generic database}

\item{queryFiles}{(list(char)) paths to query files}

\item{resDir}{(char) path to output directory}

\item{verbose}{(logical) print messages}

\item{JavaMemory}{(integer) Memory for GeneMANIA (in Gb) - a total of 
numCores*GMmemory will be used and distributed for all GM threads}

\item{numCores}{(integer) number of CPU cores for parallel processing}

\item{debugMode}{(logical) when TRUE runs jobs in serial instead of parallel and 
prints verbose messages.}

\item{debugMode}{(logical) when TRUE runs jobs in serial instead of parallel and 
prints verbose messages. Also prints system Java calls.}
}
\value{
(char) path to GeneMANIA query result files with patient similarity
rankings (*PRANK) and feature weights (*NRANK)
of results file
}
\description{
Run a query
}
\examples{
dbPath <- system.file("extdata","dbPath",package="netDx")
queryFile <- system.file("extdata","GM_query.txt",package="netDx")
runQuery(dbPath, queryFile,tempdir())
}
