% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQuery.R
\name{runQuery}
\alias{runQuery}
\title{Run a GeneMANIA query}
\usage{
runQuery(dbPath, queryFiles, resDir, verbose = TRUE, JavaMemory = 6L,
  numCores = 1)
}
\arguments{
\item{dbPath}{(char) path to directory with GeneMANIA generic database}

\item{queryFiles}{(list(char)) paths to query files}

\item{resDir}{(char) path to output directory}

\item{verbose}{(logical) print messages}

\item{JavaMemory}{(integer) Memory for GeneMANIA (in Gb) - a total of 
numCores*GMmemory will be used and distributed for all GM threads}

\item{numCores}{(integer) number of CPU cores for parallel processing}
}
\value{
path to GeneMANIA query result file
of results file
}
\description{
Run a GeneMANIA query
}
\examples{
dbPath <- sprintf("\%s/extdata/dbPath", path.package("netDx"))
GM_query <- sprintf("\%s/extdata/GM_query.txt",
	path.package("netDx"))
runQuery(dbPath, GM_query,"/tmp")
}
