% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{sim.pearscale}
\alias{sim.pearscale}
\title{various similarity functions}
\usage{
sim.pearscale(dat, K = 20, alpha = 0.5)
}
\arguments{
\item{K, }{W (doc tba)}

\item{(data.frame)}{measures by patients}
}
\description{
various similarity functions
}
