% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTestTrain.R
\name{splitTestTrain}
\alias{splitTestTrain}
\title{Split samples into training and testing set}
\usage{
splitTestTrain(pheno_DF, pctT = 0.7, setSeed = 42, verbose = FALSE)
}
\arguments{
\item{pheno_DF}{(data.frame) patient information
Must contain the following columns:
1. ID: (char) patient IDs
2. STATUS: (char) patient classes. Values not equal to \code{predClass}
will be considered as "other"
Expects rows with unique IDs}

\item{pctT}{(numeric between 0 and 1) Fraction of patients to randomly
assign to the training set. The remainder will be used for blind test 
set}

\item{setSeed}{(integer) Random number generator seed for splitting
the training and test set. If not NULL, seed will be set to this value.}

\item{verbose}{(logical) print messages}
}
\value{
(char) vector of length \code{nrow(pheno_DF)}, with values of 
"TRAIN" or "TEST". The order corresponds to pheno_DF; a patient labelled
"TRAIN" has been assigned to the training set, and one labelled "TEST"
as been assigned to the test set.
}
\description{
Split samples into training and testing set
}
\examples{
data(TCGA_mini)
x <- splitTestTrain(pheno)
}

