% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooMutationPropagation.R
\name{thresholdSmoothedMutations}
\alias{thresholdSmoothedMutations}
\title{Apply discretization to the matrix resulted from the propagation on the
sparse patient matrix}
\usage{
thresholdSmoothedMutations(
  smoothedMutProfile,
  unsmoothedMutProfile,
  nameDataset,
  n_topXmuts = c(10)
)
}
\arguments{
\item{smoothedMutProfile}{(data.frame) continous matrix of patient profiles 
resulting from applying :.,$ s/network-based propagation algorithm 
(smoothMutations_LabelProp()) on a binary somatic mutation sparse matrix.}

\item{unsmoothedMutProfile}{(data.frame) binary somatic mutation sparse 
matrix. Rownames are unique genes. Colnames are unique patients. A cell 
contains a zero or a one.}

\item{nameDataset}{(char) for titles on plot}

\item{n_topXmuts}{(numeric between 0 and 1) percent of top mutations
to keep. This function converts these to 1.0 when binarizing, so they
remain in the thresholded output matrix; other mutations are set to zero.}
}
\value{
(data.frame) binary somatic mutation matrix which sparsity has been 
decreased
}
\description{
Apply discretization to the matrix resulted from the propagation on the
sparse patient matrix
}
\details{
This function is included in the netDx use case which involves
  propagating the sparse matrix of patient's profiles to reduce its sparsity.
  This function applies discretization on the propagated matrix of patient
  profiles. It sets to 1 the genes which got the highest propagation value.
  While, the remaining genes are set to 0. This discretization is driven by
  the fact that higher is the propagation value and higher is the chance that
  the gene is involved in the patient condition and expression/mutation
  profile. On the contrary, genes which got either a medium or a low value
  are not trustable.
}
\examples{
suppressWarnings(suppressMessages(require(MultiAssayExperiment)))
require(doParallel)

# load mutation and phenotype data
genoFile <- system.file("extdata","TGCT_mutSmooth_geno.txt",package="netDx")
geno <- read.delim(genoFile,sep="\t",header=TRUE,as.is=TRUE)
phenoFile <- system.file("extdata", "TGCT_mutSmooth_pheno.txt",
			package="netDx")
pheno <- read.delim(phenoFile,sep="\t",header=TRUE,as.is=TRUE)
rownames(pheno) <- pheno$ID

# load interaction nets to smooth over
netFile <- system.file("extdata","CancerNets.txt",package="netDx")
cancerNets <- read.delim(netFile,sep="\t",header=T,as.is=T)
# smooth mutations
prop_net <- smoothMutations_LabelProp(geno,cancerNets,numCores=1L)
genoP <- thresholdSmoothedMutations(
   prop_net,geno,"TGCT_CancerNets",c(20)
  )
}
