% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{tSNEPlotter}
\alias{tSNEPlotter}
\title{Plot tSNE}
\usage{
tSNEPlotter(psn, pheno, ...)
}
\arguments{
\item{psn}{(matrix) Patient similarity network represented as adjacency
matrix (symmetric). Row and column names are patient IDs. Note that NA
values will be replaced by very small number (effectively zero).}

\item{pheno}{(data.frame) Patient labels. ID column is patient ID and 
STATUS is patient label of interest. tSNE will colour-code nodes by 
patient label.}

\item{...}{Parameters for Rtsne() function.}
}
\value{
(Rtsne) output of Rtsne call. Side effect of tSNE plot
}
\description{
Plot tSNE
}
\details{
Plots tSNE of integrated patient similarity network using Rtsne
}
