% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeQueryFile.R
\name{writeQueryFile}
\alias{writeQueryFile}
\title{Wrapper to write GeneMANIA query file}
\usage{
writeQueryFile(
  qSamps,
  incNets = "all",
  numReturn = 1L,
  outFile,
  orgName = "predictor"
)
}
\arguments{
\item{qSamps}{(char) vector of patient IDs in query}

\item{incNets}{(char) vector of networks to include in this analysis
(features/pathway names). Useful for subset-based feature selection}

\item{numReturn}{(integer) number of patients to return in ranking file}

\item{outFile}{(char) path to output file}

\item{orgName}{(char) organism name}
}
\value{
No value. Side effect of writing the query file to
\code{outFile}
}
\description{
Wrapper to write GeneMANIA query file
}
\examples{
data(pheno)
writeQueryFile(pheno$ID[seq_len(5)], 'all',nrow(pheno), 'myquery.txt')
}
