% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeWeightedNets.R
\name{writeWeightedNets}
\alias{writeWeightedNets}
\title{Write an integrated similarity network consisting of selected networks.}
\usage{
writeWeightedNets(geneFile, netInfo, netDir, keepNets, outDir,
  filterEdgeWt = 0, writeAggNet = "MAX", limitToTop = 50L,
  limitToBottom = Inf, outFileName = NULL, writeSingleNets = FALSE,
  plotEdgeDensity = FALSE, verbose = FALSE)
}
\arguments{
\item{geneFile}{(char) path to GENES.txt file created during the making
of a generic GeneMANIA database}

\item{netInfo}{(char) path to NETWORKS.txt file created during the making
of a generic GeneMANIA database. If this table contains a column marked 
"isBinary" with 1/0 indicating if a given network contains only binary 
weights, then PropBinary similarity computation will be used (see method
description).}

\item{netDir}{(char) path to directory containing interaction networks.
Note that these are networks where the node IDs have been recoded by 
GeneMANIA (e.g. 1,2,3)}

\item{keepNets}{(char or data.frame) networks to include in integrated net
If data.frame must be in "NETWORK" column,other columns will be
ignored. Mainly included as convenience so pathway scores can passed
in table format
(NETWORK), and a multiplier constant for edges in that network (WEIGHT)}

\item{outDir}{(char) path to directory where network files should be 
written}

\item{filterEdgeWt}{(numeric) keep edges with raw edge
weight strictly greater than this value. Note that "raw" refers to 
this filter being applied before the multiplier is applied.}

\item{writeAggNet}{(char, one of: [NONE|MEAN|MAX]) Aggregate the network 
1) NONE: does not write aggregate network
2) MEAN: average of weighted edges (raw x netDx score)
3) MAX: max of raw edge weight}

\item{limitToTop}{(integer) limit to top strongest connections. Set to
Inf to list all connections. Takes precedence over limitToBottom}

\item{limitToBottom}{(integer) limit to top weakest connections. Set to
Inf to list all connections. If this and limitToTop are provided, then
limitToTop takes precedence and this is ignored.}

\item{outFileName}{(char) if provided, overrides outF. Relative to outDir}

\item{writeSingleNets}{(logical) keep/delete individual recoded nets.
TRUE results in each written to its own file; FALSE does not.}

\item{plotEdgeDensity}{(logical) plot density plot of edge weights, one
per input net. Used to troubleshoot problems introduced by specific nets.
Expects a graphic device to be open already (e.g. by pdf() call)}

\item{verbose}{(logical) print messages if TRUE}
}
\value{
If an aggregated network is written (writeAggNet=TRUE),
returns the filename of the net. Else returns an empty value.
Side effect of writing one tab-delimited file per
patient network; file name is <outDir>/<network_name>.txt, and the
aggregated net, if requested.
File format is:
1) source patient (SOURCE)
2) target patient (TARGET)
3) network name (NET_NAME)
4) weight similarity for the network (WT_SIM)
}
\description{
Write an integrated similarity network consisting of selected networks.
}
