% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeWeightedNets.R
\name{writeWeightedNets}
\alias{writeWeightedNets}
\title{Write an integrated similarity network consisting of selected networks.}
\usage{
writeWeightedNets(geneFile, netInfo, netDir, keepNets, outDir,
  filterEdgeWt = 0, writeAggNet = "MAX", limitToTop = 50L,
  writeSingleNets = FALSE, verbose = FALSE)
}
\arguments{
\item{geneFile}{(char) path to GENES.txt file created during the making
of a generic GeneMANIA database}

\item{netInfo}{(char) path to NETWORKS.txt file created during the making
of a generic GeneMANIA database}

\item{netDir}{(char) path to directory containing interaction networks.
Note that these are networks where the node IDs have been recoded by 
GeneMANIA (e.g. 1,2,3)}

\item{keepNets}{(data.frame) networks to retain in the final network
(NETWORK), and a multiplier constant for edges in that network (WEIGHT)}

\item{outDir}{(char) path to directory where network files should be 
written}

\item{filterEdgeWt}{(numeric) keep edges with raw edge
weight strictly greater than this value. Note that "raw" refers to 
this filter being applied before the multiplier is applied.}

\item{writeAggNet}{(char, one of: [NONE|MEAN|MAX]) Aggregate the network 
1) NONE: does not write aggregate network
2) MEAN: average of weighted edges (raw x netDx score)
3) MAX: max of raw edge weight}

\item{limitToTop}{(integer) limit to top strongest connections. Set to
Inf to list all connections}

\item{writeSingleNets}{(logical) keep/delete individual recoded nets.
TRUE results in each written to its own file; FALSE does not.}

\item{verbose}{(logical) print messages if TRUE}
}
\value{
If an aggregated network is written (writeAggNet=TRUE),
returns the filename of the net. Else returns an empty value.
Side effect of writing one tab-delimited file per
patient network; file name is <outDir>/<network_name>.txt, and the
aggregated net, if requested.
File format is:
1) source patient (SOURCE)
2) target patient (TARGET)
3) network name (NET_NAME)
4) weight similarity for the network (WT_SIM)
}
\description{
Write an integrated similarity network consisting of selected networks.
}

